% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fcms.R
\name{aggregate_conventional_fcms}
\alias{aggregate_conventional_fcms}
\title{aggregate_conventional_fcms}
\usage{
aggregate_conventional_fcms(
  adj_matrices = list(matrix()),
  aggregation_function = c("mean", "median"),
  include_zeroes = TRUE
)
}
\arguments{
\item{adj_matrices}{A list type object of fcms. Must have a length greater
than 1.}

\item{aggregation_function}{"mean" or "median"}

\item{include_zeroes}{TRUE/FALSE Whether to include zeroes in the mean/median
calculations. (i.e. if edges not included in a map should count as a zero-weighted
edge or not at all)}
}
\value{
A single fcm calculated as the aggregate of the input adjacency matrices
}
\description{
Construct the aggregate fcm from a group (list)
of fcms. Via mean or median.
}
\details{
Construct an aggregate fcm from a group (list) of fcms

Add details here
}
