% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm.R
\name{aggregate_fcm}
\alias{aggregate_fcm}
\title{Construct an aggregate fcm from a group (list) of fcms}
\usage{
aggregate_fcm(adj_matrices = list(), aggregation_fun = "mean", IDs = c())
}
\arguments{
\item{aggregation_fun}{"mean" or "median"}

\item{IDs}{A list of names for each node (must have n items)}

\item{fcm_list}{A list type object of fcms. Must have a length greater
than 1.}
}
\value{
A single fcm calculated as the aggregate of the input adjacency matrices
}
\description{
Construct the aggregate fcm from a group (list)
of fcms. Via mean or median.
}
\details{
Add details here
}
