% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fcms.R
\name{aggregate_fcms}
\alias{aggregate_fcms}
\title{aggregate_fcms}
\usage{
aggregate_fcms(
  adj_matrices = list(matrix()),
  aggregation_function = c("mean", "median"),
  include_zeroes = FALSE
)
}
\description{
aggregate_fcms
}
