% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fcms.R
\name{aggregate_fcms_w_ivfns}
\alias{aggregate_fcms_w_ivfns}
\title{aggregate_fcms_w_ivfns}
\usage{
aggregate_fcms_w_ivfns(
  adj_matrices = list(matrix()),
  aggregation_function = c("mean", "median"),
  include_zeroes = TRUE
)
}
\arguments{
\item{adj_matrices}{A list type object of fcms. Must have a length greater
than 1.}

\item{include_zeroes}{TRUE/FALSE Whether to include zeroes in the mean/median
calculations. (i.e. if edges not included in a map should count as a zero-weighted
edge or not at all)}

\item{aggregation_fun}{"mean" or "median"}
}
\value{
A single fcm w/ edges represented as IVFNs calculated as the aggregate
of the input adjacency matrices
}
\description{
Construct an aggregate fcm w/ edges represented as IVFNs from a group (list)
of fcms w/ edges represented as IVFNs
}
