% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcmconfr.R
\name{autoplot.fcmconfr}
\alias{autoplot.fcmconfr}
\title{Autoplot fcmconfr}
\usage{
\method{autoplot}{fcmconfr}(object, ...)
}
\arguments{
\item{object}{A direct output of the \code{\link{fcmconfr}} function}

\item{...}{Additional inputs}
}
\value{
An autoplot plot of an fcmconfr object's results
}
\description{
Generate a generic plot visualizing \code{\link{fcmconfr}} results. Call the
function name directly (\code{\link{autoplot.fcmconfr}})) without parentheses
to see the exact code to generate the plots, then copy-and-paste and edit
as needed.
}
\details{
This function produces slightly different outputs for \code{\link{fcmconfr}}
outputs generated from conventional, ivfn, and tfn FCMs.
}
\examples{
NULL
}
