% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcmconfr.R
\name{build_fgcmconfr_models}
\alias{build_fgcmconfr_models}
\title{build_fgcmconfr_models}
\usage{
build_fgcmconfr_models(
  grey_adj_matrices,
  sampling,
  aggregation_fun = "mean",
  samples,
  nodes,
  parallel,
  show_progress
)
}
\arguments{
\item{grey_adj_matrices}{A list of n x n adjacencey matrices representing fcms}

\item{sampling}{The sampling method to be applied. Must be one of the following: "nonparametric", "uniform", or "triangular"}

\item{samples}{The number of samples to draw with the selected sampling method. Also,
the number of sampled models to generate}

\item{nodes}{A vector of node names (IDs) present in every adjacency matrix}

\item{parallel}{TRUE/FALSE Whether to utilize parallel processing}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}
}
\description{
This function generates n fgcm models whose edge weights are sampled from either
the defined edge values in a set of adjacency matrices or continuous (uniform or triangular)
parametric distributions derived from the sets of edge values, and stores them
as a list of adjacency matrices.
}
\details{
[ADD DETAILS HERE!!!]

Use vignette("fgcm-class") for more information.
}
