% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmcm.R
\name{build_fmcm_models_from_grey_adj_matrix}
\alias{build_fmcm_models_from_grey_adj_matrix}
\title{build_fmcm_models_from_grey_adj_matrix}
\usage{
build_fmcm_models_from_grey_adj_matrix(
  grey_adj_matrix = matrix(),
  mode_adj_matrix = matrix(),
  n_sims = 100,
  parallel = TRUE,
  distribution = "uniform",
  show_progress = TRUE,
  IDs = c()
)
}
\arguments{
\item{grey_adj_matrix}{An n x n grey adjacency matrix that represents the edge weights
of an FCM}

\item{mode_adj_matrix}{An n x n adjacency matrix that represents the mode of the
distribution of edge weights for each edge represented in the grey_adj_matrix}

\item{n_sims}{The number of simulated fcm's to generate}

\item{parallel}{TRUE/FALSE Whether to utilize parallel processing}

\item{distribution}{A statistical distribution to draw random samples from.
Must be one of the following: 'uniform', 'gaussian', 'beta', or 'triangular'}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}

\item{IDs}{A list of names for each node (must have n items)}
}
\description{
This function generates n fcm models whose edge weights are sampled from the
defined distribution ('uniform', 'gaussian', 'beta', or 'triangular') and stores
them as a list of fmcm objects
}
\details{
[ADD DETAILS HERE!!!]

Use vignette("fmcm-class") for more information.
}
