% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmcmcmr.R
\name{build_fmcmcmr_models}
\alias{build_fmcmcmr_models}
\title{build_fmcmcmr_models}
\usage{
build_fmcmcmr_models(
  adj_matrix = matrix(),
  IDs = c(),
  n_sims = 100,
  parallel = TRUE,
  distribution = "uniform",
  ...
)
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents the edge weights
of an FCM}

\item{IDs}{A list of names for each node (must have n items)}

\item{n_sims}{The number of simulated fcm's to generate}

\item{parallel}{TRUE/FALSE Whether to utilize parallel processing}

\item{distribution}{A statistical distribution to draw random samples from.
Must be one of the following: 'uniform', 'gaussian', 'beta', or 'triangular'}

\item{...}{Additional adj_matrix objects whose weights describe shape parameters
of the chosen distribution.
IF distribution = "uniform", must include: lower_adj_matrix and upper_adj_matrix objects
IF distribution = "gaussian", must include: sd_adj_matrix
IF distribution = "beta", must include: sd_adj_matrix
IF distribution = "triangular", must include: lower_adj_matrix, upper_adj_matrix, and mode_adj_matrix objects
 Note: if no mode_adj_matrix given, will assume its values are the average of the lower and upper adj_matrix objects.}
}
\description{
This function generates n fcm models whose edge weights are sampled from the
defined distribution ('uniform', 'gaussian', 'beta', or 'triangular') and stores
them as a list of fmcmcmr objects
}
\details{
[ADD DETAILS HERE!!!]

Use vignette("fmcmcmr-class") for more information.
}
