% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{build_models_from_adj_matrices_with_unconventional_values_as_distributions}
\alias{build_models_from_adj_matrices_with_unconventional_values_as_distributions}
\title{build_models_from_adj_matrices_with_unconventional_values_as_distributions}
\usage{
build_models_from_adj_matrices_with_unconventional_values_as_distributions(
  adj_matrices_with_unconventional_values_as_distributions = list(matrix()),
  aggregation_fun = c("mean", "median"),
  show_progress = TRUE
)
}
\arguments{
\item{adj_matrices_with_unconventional_values_as_distributions}{A list of
adjacency matrices with numeric and either grey_number or triangular_number elements
represented as their corresponding distributions.}

\item{aggregation_fun}{Either mean or median. The function to aggregate
distributions representing the same edges across multiple adjacency matrices}

\item{show_progress}{TRUE/FALSE Whether to display a progress bar at runtime}
}
\description{
[ADD DETAILS HERE!!!!]
}
\details{
[ADD DETAILS HERE!!!!]

Use vignette("fcmconfr-class") for more information.
}
