% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{build_monte_carlo_fcms}
\alias{build_monte_carlo_fcms}
\title{build_monte_carlo_fcms}
\usage{
build_monte_carlo_fcms(
  adj_matrix_list = list(matrix()),
  N_samples = integer(),
  include_zeroes = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{include_zeroes}{TRUE/FALSE Whether to incorporate zeroes as intentionally-defined
edge weights or ignore them in aggregation}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}

\item{unconventional_adj_matrix_list}{A list of n x n adjacencey matrices representing fcms}

\item{samples}{The number of samples to draw with the selected sampling method. Also,
the number of sampled models to generate}
}
\description{
This function generates N fcm adjacency matrices whose edge weights are sampled
from edge values (that may be numeric, grey_numbers, or triangular_numbers) and
stores them as a list of adjacency matrices.
}
\details{
If an edge is represented by multiple grey/triangular_numbers, then those distributions
are averaged together to create the aggregate distribution to sample from.

Use vignette("fcmconfr-class") for more information.
}
