% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{build_monte_carlo_fcms_from_conventional_adj_matrices}
\alias{build_monte_carlo_fcms_from_conventional_adj_matrices}
\title{build_monte_carlo_fcms_from_conventional_fcms}
\usage{
build_monte_carlo_fcms_from_conventional_adj_matrices(
  adj_matrix_list = list(Matrix::sparseMatrix()),
  N_samples = integer(),
  include_zeroes = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{include_zeroes}{TRUE/FALSE Whether to incorporate zeroes as intentionally-defined
edge weights or ignore them in aggregation}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}

\item{adj_matrices}{A list of n x n adjacencey matrices representing fcms}

\item{sampling}{The sampling method to be applied. Must be one of the following: "nonparametric", "uniform", or "triangular"}

\item{samples}{The number of samples to draw with the selected sampling method. Also,
the number of sampled models to generate}
}
\description{
This function generates n fcm models whose edge weights are sampled from either
the defined edge values in a set of adjacency matrices derived from the sets
of edge values, and stores them as a list of adjacency matrices.
}
\details{
[ADD DETAILS HERE!!!]

Use vignette("fcmconfr-class") for more information.
}
