% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{build_monte_carlo_fcms_from_fuzzy_adj_matrices}
\alias{build_monte_carlo_fcms_from_fuzzy_adj_matrices}
\title{build_monte_carlo_fcms_from_fuzzy_adj_matrices}
\usage{
build_monte_carlo_fcms_from_fuzzy_adj_matrices(
  fuzzy_adj_matrix_list = list(data.frame()),
  fuzzy_adj_matrix_list_class = c("fcm", "fgcm", "fcm_w_tfn"),
  N_samples = integer(),
  include_zeroes = FALSE,
  show_progress = TRUE
)
}
\arguments{
\item{fuzzy_adj_matrix_list}{A list of n x n fuzzy adjacencey matrices representing fcms}

\item{fuzzy_adj_matrix_list_class}{"fgcm" or "fcm_w_tfn" - the class of elements in the fuzzy_adj_matrix_list}

\item{N_samples}{The number of samples to draw from the corresponding distribution}

\item{include_zeroes}{TRUE/FALSE Whether to incorporate zeroes as intentionally-defined
edge weights or ignore them in aggregation}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}
}
\description{
This function generates n fcm adjacency matrices whose edge weights are sampled
from edge values (that may be numeric, grey_numbers, or triangular_numbers) and
stores them as a list of adjacency matrices.
}
\details{
If an edge is represented by multiple grey/triangular_numbers, then those distributions
are averaged together to create the aggregate distribution to sample from.

Use vignette("fcmconfr-class") for more information.
}
