% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{build_unconventional_fcmconfr_models}
\alias{build_unconventional_fcmconfr_models}
\title{build_unconventional_fcmconfr_models}
\usage{
build_unconventional_fcmconfr_models(
  unconventional_adj_matrix_list,
  aggregation_fun,
  samples,
  include_zeroes = FALSE,
  show_progress = TRUE
)
}
\arguments{
\item{unconventional_adj_matrix_list}{A list of n x n adjacencey matrices representing fcms}

\item{aggregation_fun}{The aggregation function, either mean or median}

\item{samples}{The number of samples to draw with the selected sampling method. Also,
the number of sampled models to generate}

\item{include_zeroes}{TRUE/FALSE Whether to incorporate zeroes as intentionally-defined
edge weights or ignore them in aggregation}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}
}
\description{
This function generates n fcm adjacency matrices whose edge weights are sampled
from edge values (that may be numeric, grey_numbers, or triangular_numbers) and
stores them as a list of adjacency matrices.
}
\details{
If an edge is represented by multiple grey/triangular_numbers, then those distributions
are averaged together to create the aggregate distribution to sample from.

Use vignette("fcmconfr-class") for more information.
}
