% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcm.R
\name{c.grey_number}
\alias{c.grey_number}
\title{c.grey_number}
\usage{
\method{c}{grey_number}(...)
}
\arguments{
\item{...}{a set of grey_number objects}
}
\description{
This forces the output of c() to the equivalent of list() only for inputs of
type grey_number
}
\details{
For grey_number objects, c() combines all of the lower and upper data into
a single grey_number object, but list() returns the expected output of a
list of distinct grey_number objects.
}
\examples{
c(grey_number(0, 1), grey_number(0.2, 0.5))
}
