% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_triangular_cognitive_maps.R
\name{c.tfn}
\alias{c.tfn}
\title{c.tfn}
\usage{
\method{c}{tfn}(...)
}
\arguments{
\item{...}{a set of tfn objects}
}
\description{
This forces the output of c() to the equivalent of list() only for inputs of
type tfn
}
\details{
For tfn objects, c() combines all of the lower and upper data into
a single tfn object, but list() returns the expected output of a
list of distinct tfn objects.
}
\examples{
c(tfn(0, 1), tfn(0.2, 0.5))
}
