% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_triangular_cognitive_maps.R
\name{c.triangular_number}
\alias{c.triangular_number}
\title{c.triangular_number}
\usage{
\method{c}{triangular_number}(...)
}
\arguments{
\item{...}{a set of triangular_number objects}
}
\description{
This forces the output of c() to the equivalent of list() only for inputs of
type triangular_number
}
\details{
For triangular_number objects, c() combines all of the lower and upper data into
a single triangular_number object, but list() returns the expected output of a
list of distinct triangular_number objects.
}
\examples{
c(triangular_number(0, 1), triangular_number(0.2, 0.5))
}
