% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcm.R
\name{calculate_greyness}
\alias{calculate_greyness}
\title{calculate_greyness}
\usage{
calculate_greyness(grey_num = grey_number(), domain = 2)
}
\arguments{
\item{grey_num}{a grey_number object}

\item{domain}{a vector c() of the minimum and maximum possible values a grey number
(or edge) may take. Either 1 if domain is bounded by c(0, 1) (i.e. only positive values)
or 2 if domain is bounded by c(-1, 1) (i.e. positive and negative values)}
}
\description{
This calculates the "greyness" of a grey number. Greyness describes uncertainty
and is defined as the grey number's range (absolute value of the upper value - lower value)
divided by the total possible range which is either 1 if edges can only contain
positive values or 2 if edges can be either positive or negative.
}
\details{
Greyness assigns a numeric value to describe the percent of the possibility space
that the grey number occupies. It is like a ratio of observed uncertainty to
maximum possible uncertainty. See https://doi.org/10.1016/j.eswa.2010.04.085

Use vignette("fgcm-class") for more information.
}
