% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_grey_cognitive_maps.R
\name{calculate_next_fgcm_state_vector}
\alias{calculate_next_fgcm_state_vector}
\title{calculate_next_fgcm_state_vector}
\usage{
calculate_next_fgcm_state_vector(
  grey_adj_matrix = matrix(),
  state_vector = c(),
  activation = "kosko",
  algorithm = "salmeron"
)
}
\arguments{
\item{grey_adj_matrix}{An n x n adjacency matrix that represents an FCM}

\item{state_vector}{A list state values at a particular iteration in an fcm simulation}

\item{activation}{The activation function to be applied. Must be one of the following:
'kosko', 'modified-kosko', or 'papageorgiou'.}

\item{algorithm}{The algorithm to run calculate the activation vector. Must
be one of the following: 'salmeron' or 'concepcion'. The 'salmeron'
algorithm is the traditional one applied in most papers (https://doi.org/10.1016%2Fj.eswa.2010.04.085),
while the 'concepcion' algorithm is a recent advancement that reduces the
calculation steps (https://doi.org/10.1007/978-3-030-52705-1_34) with a minimal reduction in accuracy}
}
\description{
This calculates the next iteration of a state vector in an fcm simulation
based on the kosko, modified-kosko, or papageorgiou activation functions
}
\details{
The state of the art of fcm typically applies one of three activation functions
in calculating iterative state vector values: kosko, modified-kosko, and
papageorgiou (as identified in Gonzales et al. 2018 - https://doi.org/10.1142/S0218213018600102).

kosko: Only considers the current iteration (Kosko, 1986 - https://doi.org/10.1016/S0020-7373(86)80040-2)

modified-kosko: The previous value of a node influences its future value (Stylio & Groumpos, 2004 - https://doi.org/10.1109/TSMCA.2003.818878)

papageorgiou: Like modified-kosko, but assigns nodes with no value with a
value of 0.5 to reduce the influence that a lack of initial state information
can have on the simulation output (Papageorgiou, 2011 - https://doi.org/10.1016/j.asoc.2009.12.010)=

Use vignette("fgcm-class") for more information.
}
