% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcm.R
\name{calculate_next_fgcm_state_vector_with_salmeron_algorithm}
\alias{calculate_next_fgcm_state_vector_with_salmeron_algorithm}
\title{calculate_next_fgcm_state_vector_with_salmeron_algorithm}
\usage{
calculate_next_fgcm_state_vector_with_salmeron_algorithm(
  state_vector = c(),
  grey_adj_matrix_col_vector = c(),
  activation = "kosko"
)
}
\arguments{
\item{state_vector}{A list state values at a particular iteration in an fcm simulation}

\item{grey_adj_matrix_col_vector}{An column of an adjacency matrix, typically as a list}

\item{activation}{The activation function to be applied. Must be one of the following:
'kosko', 'modified-kosko', or 'papageorgiou'.}
}
\description{
This calculates the next iteration of an individual node of a state vector in
an fgcm simulation using the algorithm defined by Salmeron 2010 -
https://doi.org/10.1016%2Fj.eswa.2010.04.085 and using the kosko,
modified-kosko, or papageorgiou activation functions
}
\details{
This algorithm from Salmeron 2010 - https://doi.org/10.1016%2Fj.eswa.2010.04.085

Use vignette("fgcm-class") for more information.
}
