% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_adj_matrix_list_is_list}
\alias{check_adj_matrix_list_is_list}
\title{check_adj_matrix_list_is_list}
\usage{
check_adj_matrix_list_is_list(adj_matrix_list = list(matrix()))
}
\arguments{
\item{adj_matrix_list}{A list of n x n adjacency matrix that represents an FCM}
}
\description{
Check that the input adj_matrix_list is either a list of adj. matrices or if
only one adj. matrix is given, abstract that matrix within a list (this is to
comply with lapply functionalities within the codebase)
}
\details{
This is just in case the function isn't passed an actual list of adj. matrices.
In R, both matrix and list are of class "list" so you have to do some odd checks
to check if the input is just a singular matrix or a list of matrices

Intended for developer use only to improve package readability.
}
