% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_if_local_machine_has_access_to_parallel_processing_functionalities}
\alias{check_if_local_machine_has_access_to_parallel_processing_functionalities}
\title{Check if the local machine can access internal parallel processing functionalities}
\usage{
check_if_local_machine_has_access_to_parallel_processing_functionalities(
  use_parallel,
  use_show_progress
)
}
\arguments{
\item{use_parallel}{TRUE/FALSE The user intends to use parallel processing}

\item{use_show_progress}{TRUE/FALSE The user intends to display progress bars}
}
\value{
TRUE/FALSE Whether the machine has access to the dependencies to
access internal parallel processing functionalities
}
\description{
Check whether the local machine has access to the necessary packages to
run code in parallel and/or using a progress bar. Specifically, checks for
the parallel, doSNOW, foreach, and pbapply packages.
}
\details{
Confirms that a local machine can access the required packages for parallel
processing and/or displaying progress bars at runtime. Will revise inputs
if particular packages are unavailable and warn the user of such changes, but will
not halt a run.
}
\examples{
check_if_local_machine_has_access_to_parallel_processing_functionalities(use_parallel = TRUE, use_show_progress = TRUE)
}
