% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/utils-monte_carlo_model_generation_and_simulation.R
\name{check_monte_carlo_bootstrap_inputs}
\alias{check_monte_carlo_bootstrap_inputs}
\title{Check inputs for monte carlo bootstrap analysis}
\usage{
check_monte_carlo_bootstrap_inputs(
  ci_centering_function = c("mean", "median"),
  confidence_interval = 0.95,
  num_ci_bootstraps = 1000,
  parallel = TRUE,
  n_cores = integer(),
  show_progress = TRUE
)
}
\arguments{
\item{ci_centering_function}{Estimate confidence intervals about the "mean" or "median" of
inferences from the monte carlo simulations}

\item{confidence_interval}{What are of the distribution should be bounded by the
confidence intervals? (e.g. 0.95)}

\item{num_ci_bootstraps}{Repetitions for bootstrap process, if chosen}

\item{parallel}{TRUE/FALSE Whether to perform the function using parallel processing}

\item{n_cores}{Number of cores to use in parallel processing. If no input given,
will use all available cores in the machine.}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}
}
\value{
NULL; Errors if checks fail
}
\description{
Check inputs for monte carlo bootstrap analysis
}
\examples{
NULL
}
\seealso{
Other monte-carlo-model-generation-and-simulation: 
\code{\link{build_monte_carlo_fcms}()},
\code{\link{build_monte_carlo_fcms_from_conventional_adj_matrices}()},
\code{\link{build_monte_carlo_fcms_from_fuzzy_set_adj_matrices}()},
\code{\link{check_build_monte_carlo_fcms_inputs}()},
\code{\link{check_infer_fcm_set_inputs}()},
\code{\link{get_mc_simulations_inference_CIs_w_bootstrap}()},
\code{\link{infer_fcm_set}()}
}
\concept{monte-carlo-model-generation-and-simulation}
\keyword{internal}
