% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm.R
\name{check_simulated_fcmconfr_models}
\alias{check_simulated_fcmconfr_models}
\title{check_simulated_fcmconfr_models}
\usage{
check_simulated_fcmconfr_models(
  input_adj_matrices = list(matrix()),
  sampled_adj_matrices = list(matrix()),
  sampling = c("nonparametric", "uniform", "triangular"),
  samples = integer(),
  nodes = c(),
  parallel = TRUE,
  n_cores = integer(),
  show_progress = TRUE
)
}
\arguments{
\item{input_adj_matrices}{A list of n x n adjacency matrices used as inputs to fcmconfr}

\item{sampling}{The sampling method to be applied. Must be one of the following: "nonparametric", "uniform", or "triangular"}

\item{samples}{The number of samples to draw with the selected sampling method. Also,
the number of sampled models to generate}

\item{nodes}{A vector of node names (IDs) present in every adjacency matrix}

\item{parallel}{TRUE/FALSE Whether to utilize parallel processing}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}

\item{sampled_Adj_matrices}{A list of n x n adjacency matrices constructed with build_fcmconfr_models}
}
\description{
This functions checks whether a set of sampled adj matrices are representative
of the selected distribution.
}
\details{
[ADD DETAILS HERE!!!]

Use vignette("fcm-class") for more information.
}
