% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_simulation_inputs}
\alias{check_simulation_inputs}
\title{check_simulation_inputs}
\usage{
check_simulation_inputs(
  adj_matrix,
  initial_state_vector,
  clamping_vector,
  activation,
  squashing,
  lambda,
  max_iter,
  min_error
)
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents an FCM}

\item{initial_state_vector}{A list state values at the start of an fcm simulation}

\item{clamping_vector}{A list of values representing specific actions taken to
control the behavior of an FCM. Specifically, non-zero values defined in this vector
will remain constant throughout the entire simulation as if they were "clamped" at those values.}

\item{activation}{The activation function to be applied. Must be one of the following:
'kosko', 'modified-kosko', or 'rescale'.}

\item{squashing}{A squashing function to apply. Must be one of the following:
'bivalent', 'saturation', 'trivalent', 'tanh', or 'sigmoid'.}

\item{lambda}{A numeric value that defines the steepness of the slope of the
squashing function when tanh or sigmoid are applied}

\item{max_iter}{The maximum number of iterations to run if the minimum error value is not achieved}

\item{min_error}{The lowest error (sum of the absolute value of the current state
vector minus the previous state vector) at which no more iterations are necessary
and the simulation will stop}
}
\description{
Confirm that all inputs will work with the simulation function and return
appropriate error messages where necessary
}
\details{
This checks that all inputs for a simulation function are of an appropriate
format, and also fills in missing inputs for initial_state_vector, clamping_vector,
and IDs when appropriate.
}
