% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_triangular_cognitive_maps.R
\name{clean_ftcm_simulation_output}
\alias{clean_ftcm_simulation_output}
\title{clean_tfcm_simulation_output}
\usage{
clean_ftcm_simulation_output(output_obj, IDs)
}
\arguments{
\item{output_obj}{An ftcm simulation output object}

\item{IDs}{A list of names for each node (must have n items). If empty, will use
column names of adjacancy matrix (if given).}
}
\description{
This adds quality-of-life improvements and detail to simulation output objects
such as adding column names and an iter column
}
