% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_and_simulate_fcm.R
\name{clean_simulation_output}
\alias{clean_simulation_output}
\title{clean_simulation_output}
\usage{
clean_simulation_output(output_obj, concepts)
}
\arguments{
\item{output_obj}{An fcm_w_fcm_w_tfn simulation output object}

\item{concepts}{A list of names for each node (must have n items). If empty, will use
column names of adjacancy matrix (if given).}
}
\description{
This adds quality-of-life improvements and detail to simulation output objects
such as adding column names and an iter column
}
