% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm.R
\name{confer_fcm}
\alias{confer_fcm}
\title{confer_fcm}
\usage{
confer_fcm(
  adj_matrix = matrix(),
  initial_state_vector = c(),
  clamping_vector = c(),
  activation = "kosko",
  squashing = "tanh",
  lambda = 1,
  max_iter = 100,
  min_error = 1e-05,
  lambda_optimization = "none",
  IDs = c()
)
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents an FCM}

\item{initial_state_vector}{A list state values at the start of an fcm simulation}

\item{clamping_vector}{A list of values representing specific actions taken to
control the behavior of an FCM. Specifically, non-zero values defined in this vector
will remain constant throughout the entire simulation as if they were "clamped" at those values.}

\item{activation}{The activation function to be applied. Must be one of the following:
'kosko', 'modified-kosko', or 'rescale'.}

\item{squashing}{A squashing function to apply. Must be one of the following:
'bivalent', 'saturation', 'trivalent', 'tanh', or 'sigmoid'.}

\item{lambda}{A numeric value that defines the steepness of the slope of the
squashing function when tanh or sigmoid are applied}

\item{max_iter}{The maximum number of iterations to run if the minimum error value is not achieved}

\item{min_error}{The lowest error (sum of the absolute value of the current state
vector minus the previous state vector) at which no more iterations are necessary
and the simulation will stop}

\item{lambda_optimization}{A lambda optimization procedure to apply. Must be one
of the following: 'none' or 'koutsellis'}

\item{IDs}{A list of names for each node (must have n items). If empty, will use
column names of adjacancy matrix (if given).}
}
\description{
This confers with a baseline simulation of an FCM and a scenario (scenario vector)
to estimate how outputs change compared to the structural or expected behavior
of the system.
}
\details{
This function performs two fcm simulations and compares the output between the two.
The first simulation considers the baseline activity where no nodes are "clamped" and the
system behaves without any outside inputs. The second simulation considers a scenario where
one or multiple nodes are "clamped" so that the system is reactive to additional inputs.
The function returns the difference in simulation results between the scenario and baseline
activity to understand how system manipulations compare to structural expectations of the system.

This function produces the same output as mental modeler for the following inputs:
 - initial_state_vector = c(1, 1, ..., 1)
 - activation = "kosko"
 - squashing = either "sigmoid" or "tanh"
 - lambda = 1

Use vignette("fcm-class") for more information about each of these
functions/algorithms alongside their originating sources.
}
