% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcm.R
\name{confer_fgcm}
\alias{confer_fgcm}
\title{confer_fgcm}
\usage{
confer_fgcm(
  grey_adj_matrix = matrix(),
  activation_vector = c(),
  scenario_vector = c(),
  activation = "modified-kosko",
  squashing = "sigmoid",
  lambda = 1,
  max_iter = 50,
  min_error = 1e-05,
  IDs = c(),
  algorithm = "salmeron"
)
}
\arguments{
\item{grey_adj_matrix}{An n x n grey_adjacency matrix that represents an FCM}

\item{activation_vector}{A list state values at the start of an fcm simulation}

\item{scenario_vector}{A list of values representing specific actions taken to
control the behavior of an FCM. Specifically, non-zero values defined in this vector
will remain constant throughout the entire simulation as if they were "clamped" at those values.}

\item{activation}{The activation function to be applied. Must be one of the following:
'kosko', 'modified-kosko', or 'papageorgiou'.}

\item{squashing}{A squashing function to apply. Must be one of the following:
'bivalent', 'saturation', 'trivalent', 'tanh', or 'sigmoid'.}

\item{lambda}{A numeric value that defines the steepness of the slope of the
squashing function when tanh or sigmoid are applied}

\item{max_iter}{The maximum number of iterations to run if the minimum error value is not achieved}

\item{min_error}{The lowest error (sum of the absolute value of the current state
vector minus the previous state vector) at which no more iterations are necessary
and the simulation will stop}

\item{IDs}{A list of names for each node (must have n items). If empty, will use
column names of adjacancy matrix (if given).}

\item{algorithm}{The algorithm to run calculate the activation vector. Must
be one of the following: 'salmeron' or 'concepcion'. The 'salmeron'
algorithm is the traditional one applied in most papers (https://doi.org/10.1016%2Fj.eswa.2010.04.085),
while the 'concepcion' algorithm is a recent advancement that reduces the
calculation steps with a minimal reduction in accuracy (https://doi.org/10.1007/978-3-030-52705-1_34)}
}
\description{
This confers with a baseline simulation of an FGCM and a scenario (scenario vector)
to estimate how outputs change compared to the structural or expected behavior
of the system.
}
\details{
This function performs two fgcm simulations and compares the output between the two.
The first simulation considers the baseline activity where no nodes are "clamped" and the
system behaves without any outside inputs. The second simulation considers a scenario where
one or multiple nodes are "clamped" so that the system is reactive to additional inputs.
The function returns the difference in simulation results between the scenario and baseline
activity to understand how system manipulations compare to structural expectations of the system.

Use vignette("fgcm-class") for more information about each of these
functions/algorithms alongside their originating sources.
}
