% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{confirm_adj_matrices_have_same_concepts}
\alias{confirm_adj_matrices_have_same_concepts}
\title{confirm_adj_matrices_have_same_concepts}
\usage{
confirm_adj_matrices_have_same_concepts(
  list_of_concepts_by_adj_matrix = list()
)
}
\arguments{
\item{list_of_concepts_by_adj_matrix}{A list of lists of column names of adjacency matrices}
}
\description{
This checks a list of lists of column names of adjacency matrices and
confirms that each list is identical.
}
\details{
Note: This function does NOT take the raw adjacency matrices as an input. Rather,
it takes a list of lists of column names of adjacency matrices.
}
