% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{confirm_adj_matrices_have_same_dimensions}
\alias{confirm_adj_matrices_have_same_dimensions}
\title{confirm_adj_matrices_have_same_dimensions}
\usage{
confirm_adj_matrices_have_same_dimensions(adj_matrices = list(matrix()))
}
\arguments{
\item{adj_matrices}{A list of n x n adjacency matrices}
}
\description{
This checks that all adjacency matrices in a list have the same dimensions
(i.e. all are n x n)
}
\details{
Note: This function DOES take the raw adjacency matrices as an input.
}
