% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{confirm_adj_matrix_is_square}
\alias{confirm_adj_matrix_is_square}
\title{confirm_adj_matrix_is_square}
\usage{
confirm_adj_matrix_is_square(adj_matrix = matrix())
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents an FCM}
}
\description{
Confirm that an adjacency matrix is square (n x n)
}
\details{
Boolean. TRUE if the dimensions of an adjacency matrix are equivalent (n x n).
FALSE if not.

Intended for developer use only to improve package readability.
}
