% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{confirm_input_vector_is_compatible_with_adj_matrices}
\alias{confirm_input_vector_is_compatible_with_adj_matrices}
\title{confirm_input_vector_is_compatible_with_adj_matrices}
\usage{
confirm_input_vector_is_compatible_with_adj_matrices(
  representative_adj_matrix = matrix(),
  input_vector = c(),
  fcm_class = c("fcm", "fgcm", "fcm_w_tfn")
)
}
\arguments{
\item{representative_adj_matrix}{An adjacency matrix whose format (i.e. dims and data.types)
are representative of a larger list of adjacency matrices}

\item{input_vector}{An input vector, either the initial_state_vector input or
the clamping_vector input}

\item{fcm_class}{The class of fcm represented by the representative_adj_matrix}
}
\description{
Check whether an input vector (initial_state_vector or clamping_vector) is
compatible with the data types present in the representative_adj_matrix.
}
\details{
[ADD DETAILS HERE!!!]

Intended for developer use only to improve package readability.
}
