% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_fuzzy_numbers.R
\name{confirm_input_vector_is_compatible_with_fcm_w_tfn_adj_matrix}
\alias{confirm_input_vector_is_compatible_with_fcm_w_tfn_adj_matrix}
\title{confirm_input_vector_is_compatible_with_fcm_w_tfn_adj_matrix}
\usage{
confirm_input_vector_is_compatible_with_fcm_w_tfn_adj_matrix(
  fcm_w_tfn_adj_matrix = matrix(),
  initial_state_vector = c()
)
}
\arguments{
\item{fcm_w_tfn_adj_matrix}{An n x n triangular adjacency matrix that represents an FCM}

\item{initial_state_vector}{An n-length list of the initial states of each node in an fcm simulation}
}
\description{
Confirm that an initial state vector is algorithmically compatible with a triangular adjacency matrix
}
\details{
Boolean. TRUE if the number of entries in the initial
state vector match the number of rows/columns in the adjacency matrix and 2. The
datatypes stored within each object are the same (i.e. "numeric" vs "tfn"),
FALSE if not

Intended for developer use only to improve package readability.
}
