% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{confirm_only_numeric_data_in_adj_matrix}
\alias{confirm_only_numeric_data_in_adj_matrix}
\title{confirm_only_numeric_data_in_adj_matrix}
\usage{
confirm_only_numeric_data_in_adj_matrix(adj_matrix = matrix())
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents an FCM}
}
\description{
Confirm all values in an adj_matrix object are of type numeric
}
\details{
Check that all values in an adjacency matrix are of type numeric (i.e. int,
double, etc.)

Intended for developer use only to improve package readability.
}
