% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcm.R
\name{confirm_scenario_vector_is_compatible_with_grey_adj_matrix}
\alias{confirm_scenario_vector_is_compatible_with_grey_adj_matrix}
\title{confirm_scenario_vector_is_compatible_with_grey_adj_matrix}
\usage{
confirm_scenario_vector_is_compatible_with_grey_adj_matrix(
  grey_adj_matrix = matrix(),
  scenario_vector = c()
)
}
\arguments{
\item{grey_adj_matrix}{An n x n grey adjacency matrix that represents an FCM}

\item{scenario_vector}{An n-length list of the initial states of each node in an fcm simulation}
}
\description{
Confirm that a scenario_vector is algorithmically compatible with a grey adjacency matrix
}
\details{
Boolean. TRUE if the number of entries in the initial
state vector match the number of rows/columns in the adjacency matrix and 2. The
datatypes stored within each object are the same (i.e. "numeric" vs "grey_number"),
FALSE if not

Intended for developer use only to improve package readability.
}
