% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{confirm_unique_datatype_in_object}
\alias{confirm_unique_datatype_in_object}
\title{confirm_unique_datatype_in_object}
\usage{
confirm_unique_datatype_in_object(object, datatype = "numeric")
}
\arguments{
\item{object}{A list-like object (matrix, data.frame, list, etc.)}

\item{datatype}{The datatype of which the class of every value within the object should match}
}
\description{
Confirm that an object contains data of a single class
}
\details{
Boolean. TRUE if an object contains only data of a single class (e.g. "numeric"),
FALSE if multiple classes detected

Intended for developer use only to improve package readability.
}
