% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_and_simulate_fcm.R
\name{convert_element_to_ivfn_or_tfn_if_numeric}
\alias{convert_element_to_ivfn_or_tfn_if_numeric}
\title{convert_element_to_ivfn_or_tfn_if_numeric}
\usage{
convert_element_to_ivfn_or_tfn_if_numeric(
  element,
  desired_class = c("ivfn", "tfn")
)
}
\arguments{
\item{element}{An element in a matrix}
}
\description{
This checks whether the input element is an ordinary number or a triangular number.
If it is an ivfn or tfn, it returns the input, but if it is a numeric type
object (ordinary number), it will convert that number into an ivfn or tfn
}
