% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_and_simulate_fcm.R
\name{defuzz}
\alias{defuzz}
\title{defuzz}
\usage{
defuzz(fuzzy_number)
}
\arguments{
\item{fuzzy_number}{A fuzzy number object. Either an ivfn or tfn}
}
\description{
Convert a fuzzy number to a crisp value. For IVFNs, return the average of the
upper and lower bounds. For TFNs, return the average of the lower bound, the
mode, and the upper bound.
}
