% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_and_simulate_fcm.R
\name{defuzz_ivfn_or_tfn}
\alias{defuzz_ivfn_or_tfn}
\title{Defuzz (IVFN or TFN)}
\usage{
defuzz_ivfn_or_tfn(fuzzy_number)
}
\arguments{
\item{fuzzy_number}{A fuzzy number object. Either an ivfn or tfn}
}
\value{
A crisp value representative of the input IVFN or TFN
}
\description{
Convert a fuzzy number to a crisp value. For IVFNs, return the average of the
upper and lower bounds. For TFNs, return the average of the lower bound, the
mode, and the upper bound.
}
\examples{
defuzz_ivfn_or_tfn(ivfn(-1, 1))
defuzz_ivfn_or_tfn(tfn(-1, 0, 1))
}
