% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_fuzzy_numbers.R
\name{equalize_fcm_w_tfn_simulation_state_vectors_dataframes}
\alias{equalize_fcm_w_tfn_simulation_state_vectors_dataframes}
\title{equalize_fcm_w_tfn_simulation_state_vectors_dataframes}
\usage{
equalize_fcm_w_tfn_simulation_state_vectors_dataframes(baseline, scenario)
}
\arguments{
\item{baseline}{The output of a fcm_w_fcm_w_tfn simulation using the baseline inputs}

\item{scenario}{The output of a fcm_w_fcm_w_tfn simulation using the scenario inputs}
}
\description{
This reformats the dataframes of the simulation with the least number of iterations
(rows) and repeats the final row until dataframes across the baseline and scenario
simulations have the same number of iterations (rows)

Use vignette("fcm_w_tfn-class") for more information about each of these
functions/algorithms alongside their originating sources.
}
