% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{estimate_lambda}
\alias{estimate_lambda}
\title{Estimate lambda}
\usage{
estimate_lambda(fcm_adj_matrix = matrix(), squashing = c("sigmoid", "tanh"))
}
\arguments{
\item{squashing}{A squashing function to apply. Must be one of the following: 'tanh', or 'sigmoid'.}

\item{adj_matrix}{An n x n adjacency matrix that represents an FCM}
}
\description{
This calculates optimum lambda value for the sigmoid and tanh squashing
function that guarantees convergence of the simulation

It estimates lambda such that the 'squashed' values will be contained within
the near-linear region of the sigmoid or tanh function, and then re-normalizes
those values back to the total possibility spaces of those functions ([0, 1]
and [-1, 1] respectively).
}
\details{
This algorithm was first explored in Kottas et al. 2010 (https://doi.org/10.1007/978-3-642-03220-2_5),
expanded upon in Koutsellis et al. 2022 (https://doi.org/10.1007/s12351-022-00717-x), and
further further developed in Koutsellis et al. 2022 (https://doi.org/10.1109/IISA56318.2022.9904369).

This applies an algorithm to optimize lambda. Currently, the author only
identifies one such algorithm, but generalizes the function to leave flexibility
for the addition of newly-discovered algorithms in the future.
}
