% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmr.R
\name{fcm}
\alias{fcm}
\title{fcm (fuzzy cognitive map) S3 class}
\usage{
fcm(adj_matrix = matrix(), IDs = c())
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents an FCM}

\item{IDs}{A list of names for each node (must have n items)}
}
\description{
This class is an organization scheme for ordinary fuzzy cognitive maps (See
Kosko, XXXX for example). It stores the nodes of an FCM and its
corresponding adjacency matrix and edgelist.
}
\details{
fcm stores fcm data in forms useful to data manipulation, particular
regarding pairing with popular network analysis libraries like igraph
or visNetwork.

Use vignette("fcm-class") for more information.
}
\examples{
fcm(adj_matrix = matrix(data = c(0, 1, 1, 0), nrow = 2, ncol = 2))
}
