% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmconfr_gui.R
\name{fcmconfr_gui}
\alias{fcmconfr_gui}
\title{fcmconfr_gui}
\usage{
fcmconfr_gui()
}
\value{
On exit, this function outputs a copy-and-paste-able sample script
to call \code{\link{fcmconfr}} with the selected inputs.
}
\description{
This function opens a shiny application window designed to help users
determine inputs for the \code{\link{fcmconfr}} function.
}
\examples{
NULL
}
