% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmconfr.R
\name{fcmconfr_plot}
\alias{fcmconfr_plot}
\title{fcmconfr_plot}
\usage{
fcmconfr_plot(
  fcmconfr_output,
  swap_axes = FALSE,
  aggregate_point_fill = "#fb0009",
  monte_carlo_col_fill = "#fcdbd0",
  monte_carlo_col_alpha = 0.6
)
}
\arguments{
\item{fcmconfr_output}{Output directly from fcmconfr}

\item{swap_axes}{Switch x and y axes. Preferable when concepts have longer names}

\item{aggregate_point_fill}{Hex code value for the aggregate point fill}

\item{monte_carlo_col_fill}{Hex code value for the columns/bars representing monte carlo data}

\item{monte_carlo_col_alpha}{Between 0 and 1, the opacity of the monte_carlo_col_fill}
}
\value{
A plot of the results generated from fcmconfr
}
\description{
fcmconfr_plot
}
