% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{fcmr}
\alias{fcmr}
\title{fcmr (fuzzy cognitive map) S3 class}
\usage{
fcmr(adj_matrix = matrix())
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents an FCM}
}
\description{
This class is an organization scheme for ordinary fuzzy cognitive maps (See
Kosko, XXXX for example). It stores the nodes of an FCM and its
corresponding adjacency matrix and edgelist.
}
\details{
fcmr stores fcm data in forms useful to data manipulation, particular
regarding pairing with popular network analysis libraries like igraph
or visNetwork.

fcmr are constructed using vctrs::new_vctr instead of the typical call to
structure() because ...

Use vignette("fcmr-class") for more information.
}
\examples{
fcmr(adj_matrix = matrix(data = c(0, 1, 1, 0), nrow = 2, ncol = 2))
}
