% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcmr.R
\name{fgcmr}
\alias{fgcmr}
\title{fgcmr (fuzzy grey cognitive map) S3 class}
\usage{
fgcmr(grey_adj_matrix = matrix(), IDs = c())
}
\arguments{
\item{grey_adj_matrix}{An n x n adjacency matrix that represents an FCM}

\item{IDs}{A list of names for each node (must have n items)}
}
\description{
This class is an organization scheme for fuzzy grey cognitive maps (See
____, XXXX for example). It stores the nodes of an FGCM and its
corresponding adjacency matrix and edgelist.
}
\details{
fgcmr stores fgcm data in forms useful to data manipulation, particular
regarding pairing with popular network analysis libraries like igraph
or visNetwork.

fgcmr allow for the depiction of uncertainty in fcm as edge weights are
represented as Grey Numbers (___, XXXX). Grey Numbers represent a range of
possible values that an edge may be within.

fgcmr are constructed using vctrs::new_vctr instead of the typical call to
structure() because ...

Use vignette("fgcmr-class") for more information.
}
\examples{
NULL
}
