% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_adj_matrix_from_edgelist}
\alias{get_adj_matrix_from_edgelist}
\title{get_adj_matrix_from_edgelist}
\usage{
get_adj_matrix_from_edgelist(
  edgelist = matrix(),
  source_colname = "source",
  target_colname = "target",
  value_colname = "weight"
)
}
\arguments{
\item{edgelist}{An edgelist representing an fcm. Default column names are
"source", "target", and "weight", but these may be defined explicitly.}

\item{source_colname}{Column name in the input eddgelist that represents
edge source nodes}

\item{target_colname}{Column name in the input edgelist that represents
edge target nodes}

\item{value_colname}{Colname in th einput edgelist that represents represents
the values displayed in the adjacency matrix (i.e. weight, standard_deviation)}
}
\description{
This "gets" an adjacency matrix representing a graph described by the input
edgelist.
}
\details{
The input edgelist must have the following column names: 'source' or 'from',
'target' or 'to'. The user must manually note if different names are used
for the edgelist. An additional column may be selected to describe a value
attributed to a given edge.

The input edgelist can be either a matrix, data.frame, tibble, or
data.table type object.
}
