% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_adj_matrix_list_input_type}
\alias{get_adj_matrix_list_input_type}
\title{get_adj_matrix_list_object_type}
\usage{
get_adj_matrix_list_input_type(adj_matrix_list_input)
}
\arguments{
\item{adj_matrix_list_object}{A list of adj matrices or an individual adj matrix}
}
\description{
A short description...
}
