% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmcmcmr.R
\name{get_beta_distribution_of_values}
\alias{get_beta_distribution_of_values}
\title{get_beta_distribution_of_values}
\usage{
get_beta_distribution_of_values(mu = double(), sd = double(), n = 1000)
}
\arguments{
\item{mu}{mean of the population distribution}

\item{sd}{standard deviation of the population distribution}

\item{n}{number of samples to draw from the defined beta distribution}
}
\description{
This pulls n samples from a beta distribution described by shape parameters
defined by a mean and standard deviation
}
\details{
Use vignette("fcmcmrr-class") for more information.
}
