% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_class_of_adj_matrix}
\alias{get_class_of_adj_matrix}
\title{get_class_of_adj_matrix}
\usage{
get_class_of_adj_matrix(adj_matrix = matrix())
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents an FCM, FGCM, or FTCM}
}
\description{
Get the class of map from the input adj matrix
}
\details{
This returns the class of fcm represented by the input adjacency matrices i.e.
fcm, fgcm, ftcm, etc.

Intended for developer use only to improve package readability.
}
