% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcm.R
\name{get_domain_of_grey_adj_matrix}
\alias{get_domain_of_grey_adj_matrix}
\title{calculate_greyness}
\usage{
get_domain_of_grey_adj_matrix(grey_adj_matrix = matrix())
}
\arguments{
\item{grey_adj_matrix}{an adjacency matrix of grey_number class objects}
}
\description{
This calculates the domain of a grey adjacency matrix. The domain is the
absolute value of the difference between the maximum and minimum possible values
that a grey number may encompass within a grey adjacency matrix.
}
\details{
The domain of a grey adjacency matrix may be either 1 if it contains only positive
values (i.e. c(0, 1)) or 2 if it contains positive and negative values (i.e. c(-1, 1))
See https://doi.org/10.1016/j.eswa.2010.04.085

Use vignette("fgcm-class") for more information.
}
