% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_grey_cognitive_maps.R
\name{get_edgelist_from_grey_adj_matrix}
\alias{get_edgelist_from_grey_adj_matrix}
\title{get_edgelist_from_grey_adj_matrix}
\usage{
get_edgelist_from_grey_adj_matrix(grey_adj_matrix = matrix(), IDs = c())
}
\arguments{
\item{grey_adj_matrix}{An n x n adjacency matrix that represents an FCM}

\item{IDs}{A list of names for each node (must have n items)}
}
\description{
This "gets" an edgelist representing a graph described by the input grey adjacency
matrix.
}
\details{
The input adjacency matrix must be a square n x n matrix. It can be either
a matrix, data.frame, tibble, or data.table type object.

If the input matrix has named columns, those names will be used as node IDs
in the edgelist. Otherwise, generic node IDs will be used (1, 2, ... n)

The edgelist returns the following columns: source, target, weight
}
\examples{
NULL
}
