% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_fcm_class_from_adj_matrices}
\alias{get_fcm_class_from_adj_matrices}
\title{get_fcm_class_from_adj_matrices}
\usage{
get_fcm_class_from_adj_matrices(adj_matrices = list(matrix()))
}
\arguments{
\item{adj_matrices}{A list of n x n adjacency matrix that represents an FCM}
}
\description{
Get the class of fcm from the input adj matrices
}
\details{
This returns the class of fcm represented by the input adjacency matrices i.e.
fcm, fgcm, ftcm, etc.

Intended for developer use only to improve package readability.
}
