% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcmr.R
\name{get_grey_adj_matrix_from_list_of_grey_numbers}
\alias{get_grey_adj_matrix_from_list_of_grey_numbers}
\title{get_grey_adj_matrix_from_list_of_grey_numbers}
\usage{
get_grey_adj_matrix_from_list_of_grey_numbers(
  values = c(),
  locs = c("row-col"),
  size = numeric()
)
}
\arguments{
\item{values}{A list of grey_number or numeric type objects}

\item{locs}{A list of matrix locations/indexes (row-col) to place entries
in the values list}

\item{size}{The square (n x n) dimensions of the output matrix. Must be
greater than or equal to the number of values and locs given. Size
is assumed to be the square root of the length of the number of
values given, rounded up to the nearest whole number.}
}
\description{
This "gets" a Grey adjacency matrix from a list of input values and a list
of matrix indexes that identify where those values are located in the matrix
(row-col).
}
\details{
The input values and locations (locs) lists must have equivalent lengths, and
must be of type list.

If the input lists have named columns, those names will be carried over
in the grey adjacency matrix. Otherwise, generic node IDs will be used
(C1, C2, ... Cn).

#' Use vignette("fgcmr-class") for more information.
}
\examples{
get_grey_adj_matrix_from_list_of_grey_numbers(
 values = c(grey_number(-1, 1), grey_number(-0.5, 0.5)),
 locs = c("1-2", "2-1"),
 size = 2
)
}
