% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{get_mc_simulations_inference_CIs_w_bootstrap}
\alias{get_mc_simulations_inference_CIs_w_bootstrap}
\title{Calculate Inferences (w/ Confidence Intervals via Bootstrap) of MC FCM Simulations}
\usage{
get_mc_simulations_inference_CIs_w_bootstrap(
  mc_simulations_inference_df = data.frame(),
  inference_function = "mean",
  confidence_interval = 0.95,
  bootstrap_reps = 1000,
  bootstrap_draws_per_rep = 1000,
  parallel = TRUE,
  n_cores = integer(),
  show_progress = TRUE
)
}
\arguments{
\item{mc_simulations_inference_df}{The final values of a set of fcm simulations; also the inference of a infer_fmcm object}

\item{inference_function}{Estimate confidence intervals about the "mean" or "median" of
inferences from the monte carlo simulations}

\item{confidence_interval}{What are of the distribution should be bounded by the
confidence intervals? (e.g. 0.95)}

\item{bootstrap_reps}{Repetitions for bootstrap process, if chosen}

\item{bootstrap_draws_per_rep}{Number of samples to draw (with replacement) from
the data per bootstrap_rep}

\item{parallel}{TRUE/FALSE Whether to perform the function using parallel processing}

\item{n_cores}{Number of cores to use in parallel processing. If no input given,
will use all available cores in the machine.}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}
}
\description{
This gets the mean of the distribution of simulated values
across a given iter. Also returns the bootstrapped mean of means of the
distribution of simulated values across a given iter if called.
}
\details{
This function is designed to streamline the process of getting the mean or bootstrapped
mean of means of a distribution of simulated values across individual iterations. Use get_bootstrapped_means
to estimate the confidence intervals for the mean value across simulations.

The show_progress and parallel inputs change the functions called, but do NOT
change the output! These are allowed to be toggled on/off to increase user
control at runtime.
}
\seealso{
Other monte-carlo-model-generation-and-simulation: 
\code{\link{build_monte_carlo_fcms}()},
\code{\link{build_monte_carlo_fcms_from_conventional_adj_matrices}()},
\code{\link{build_monte_carlo_fcms_from_fuzzy_set_adj_matrices}()},
\code{\link{infer_monte_carlo_fcm_set}()},
\code{\link{monte_carlo_bootstrap_checks}()}
}
\concept{monte-carlo-model-generation-and-simulation}
