% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmccm.R
\name{get_means_of_fmccm_inferences}
\alias{get_means_of_fmccm_inferences}
\title{get_means_of_fmccm_inferences}
\usage{
get_means_of_fmccm_inferences(
  fmccm_final_states_across_sims = list(),
  get_bootstrapped_means = TRUE,
  confidence_interval = 0.95,
  bootstrap_reps = 1000,
  bootstrap_samples_per_rep = 1000,
  parallel = FALSE,
  n_cores = integer(),
  show_progress = TRUE
)
}
\arguments{
\item{fmccm_final_states_across_sims}{The final values of a set of fcm simulations; also the inferences of a confer_fmccm object}

\item{get_bootstrapped_means}{TRUE/FALSE Whether to perform bootstrap sampling to obtain
confidence intervals for the estimation of the mean value across simulations}

\item{confidence_interval}{What are of the distribution should be bounded by the
confidence intervals? (e.g. 0.95)}

\item{bootstrap_reps}{Repetitions for bootstrap process, if chosen}

\item{bootstrap_samples_per_rep}{Number of samples to draw (with replacement) from
the data per bootstrap_rep}

\item{parallel}{TRUE/FALSE Whether to perform the function using parallel processing}

\item{n_cores}{Number of cores to use in parallel processing. If no input given,
will use all available cores in the machine.}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmccm. Uses pbmapply
from the pbapply package as the underlying function.}
}
\description{
This gets the mean of the distribution of simulated values
across a given iter. Also returns the bootstrapped mean of means of the
distribution of simulated values across a given iter if called.
}
\details{
This function is designed to streamline the process of getting the mean or bootstrapped
mean of means of a distribution of simulated values across individual iterations. Use get_bootstrapped_means
to estimate the confidence intervals for the mean value across simulations.

Use vignette("fmccm-class") for more information.
}
