% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmccm.R
\name{get_quantile_of_fmccm_state_vectors_at_iter}
\alias{get_quantile_of_fmccm_state_vectors_at_iter}
\title{get_quantile_of_fmccm_state_vectors_at_iter}
\usage{
get_quantile_of_fmccm_state_vectors_at_iter(
  fmccm_state_vectors_by_iter = list(),
  quantile = 0.5
)
}
\arguments{
\item{fmccm_state_vectors_by_iter}{Output of get_simulated_values_across_iters}

\item{quantile}{The quantile to return. see ?quantile() for more}
}
\description{
This gets the user-input quantile of the distribution of simulated values
across a given iter, or all iters
}
\details{
This function is designed to streamline the process of getting the custom quantiles
of a distribution of simulated values across an individual iteration.

Use vignette("fmccm-class") for more information.
}
