% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmcm.R
\name{get_quantile_of_fmcm_inferences}
\alias{get_quantile_of_fmcm_inferences}
\title{get_quantile_of_fmcm_state_vectors_at_iter}
\usage{
get_quantile_of_fmcm_inferences(fmcm_inferences = data.frame(), quantile = 0.5)
}
\arguments{
\item{fmcm_inferences}{Output of get_simulated_values_across_iters}

\item{quantile}{The quantile to return. see ?quantile() for more}
}
\description{
This gets the user-input quantile of the distribution of simulated values
across a given iter, or all iters
}
\details{
This function is designed to streamline the process of getting the custom quantiles
of a distribution of simulated values across an individual iteration.

Use vignette("fmcm-class") for more information.
}
