% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmcmcmr.R
\name{get_quantiles_of_simulated_values_across_iters}
\alias{get_quantiles_of_simulated_values_across_iters}
\title{get_quantiles_of_simulated_values_across_iters}
\usage{
get_quantiles_of_simulated_values_across_iters(
  simulated_values_across_iters = list(),
  lower_quantile = 0.025,
  upper_quantile = 0.975,
  get_bootstrapped_means = FALSE,
  bootstrap_reps = 1000,
  bootstrap_samples_per_rep = 1000
)
}
\arguments{
\item{simulated_values_across_iters}{Output of get_simulated_values_across_iters}

\item{lower_quantile}{The lower quantile. see ?quantile() for more}

\item{upper_quantile}{The upper quantile, see ?quantile() for more}

\item{get_bootstrapped_means}{TRUE/FALSE Whether to perform bootstrap sampling to obtain
confidence intervals for the estimation of the mean value across simulations}

\item{bootstrap_reps}{Repetitions for bootstrap process, if chosen}

\item{bootstrap_samples_per_rep}{Number of samples to draw (with replacement) from
the data per bootstrap_rep}
}
\description{
This gets the user-input quantiles of the distribution of simulated values
across a given iter
}
\details{
This function is designed to streamline the process of getting the custom quantiles
of a distribution of simulated values across an individual iteration. Use get_bootstrapped_means
to estimate the confidence intervals for the mean value across simulations.

Use vignette("fmcmcmr-class") for more information.
}
