% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmcmcmr.R
\name{get_simulated_values_across_iters}
\alias{get_simulated_values_across_iters}
\title{get_simulated_values_across_iters}
\usage{
get_simulated_values_across_iters(fmcmcmr_simulation, iter = integer())
}
\arguments{
\item{fmcmcmr_simulation}{Output from simulate_fmcmcmr_models, represents results across
each simulated fcm}

\item{iter}{The iteration at which to return a distribution of values i.e. get
values at iteration 'iter'. If no value is given, returns distributions of values across
each iter in the data set}
}
\description{
This outputs simulation results at a specific or all iters
}
\details{
This function is designed to streamline the process of getting a distribution
of values at a specific iter to create histograms or perform bootstrap sampling
operations to estimate confidence boundes.

Use vignette("fmcmcmr-class") for more information.
}
