% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmccm.R
\name{get_state_vectors_by_iter_from_fmccm_sims}
\alias{get_state_vectors_by_iter_from_fmccm_sims}
\title{get_state_vectors_by_iter_from_fmccm_sims}
\usage{
get_state_vectors_by_iter_from_fmccm_sims(
  state_vectors_by_sim,
  iter = integer()
)
}
\arguments{
\item{state_vectors_by_sim}{Output from simulate_fmccm_models, represents results across
each simulated fcm}

\item{iter}{The iteration at which to return a distribution of values i.e. get
values at iteration 'iter'. If no value is given, returns distributions of values across
each iter in the data set}
}
\description{
This outputs simulation results at a specific or all iters
}
\details{
This function is designed to streamline the process of getting a distribution
of values at a specific iter to create histograms or perform bootstrap sampling
operations to estimate confidence boundes.

Use vignette("fmccm-class") for more information.
}
