% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_triangular_cognitive_maps.R
\name{get_triangular_adj_matrix_from_lower_mode_and_upper_adj_matrices}
\alias{get_triangular_adj_matrix_from_lower_mode_and_upper_adj_matrices}
\title{get_triangular_adj_matrix_from_lower_mode_and_upper_adj_matrices}
\usage{
get_triangular_adj_matrix_from_lower_mode_and_upper_adj_matrices(
  lower = matrix(),
  mode = matrix(),
  upper = matrix()
)
}
\arguments{
\item{lower}{An n x n adjacency matrix that represents the lower limits of
edges in an FCM}

\item{mode}{An n x n adjacency matrix that represents the modes of edges in an FCM}

\item{upper}{An n x n adjacency matrix that represents the upper limits of
edges in an FCM}
}
\description{
This "gets" a triangular adjacency matrix from an adjacency matrix of the lower
limits of edges in an FCM and an adjacency matrix of the upper limits of edges
in an FCM.
}
\details{
The input adjacency matrices must square n x n matrices with the same dimensions.
The input can be either matrix, data.table, tibble, or data.table type objects,
but the output will always be a data.frame. This is for output readability.
data.table and tibble objects work logically, but their outputs require
additional steps to parse from the user's perspective.

If the input matrices have named columns, those names will be carried over
in the triangular adjacency matrix. Otherwise, generic node IDs will be used
(C1, C2, ... Cn).

#' Use vignette("ftcm-class") for more information.
}
\examples{
get_triangular_adj_matrix_from_lower_mode_and_upper_adj_matrices(
 lower = matrix(data = c(0, 0.2, 0, 0.5), nrow = 2, ncol = 2),
 mode = matrix(data = c(0, 0.3, 0, 0.6), nrow = 2, ncol = 2),
 upper = matrix(data = c(0, 0.4, 0, 0.7), nrow = 2, ncol = 2)
)
}
