% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmcm.R
\name{get_triangular_distribution_of_values}
\alias{get_triangular_distribution_of_values}
\title{get_triangular_distribution_of_values}
\usage{
get_triangular_distribution_of_values(
  lower = double(),
  upper = double(),
  mode = double(),
  n = 1000
)
}
\arguments{
\item{lower}{lower limit or minimum of the sample space}

\item{upper}{upper limit or maximum of the sample space}

\item{mode}{peak of the sample space}

\item{n}{number of samples to draw from the triangular distribution}
}
\description{
This pulls n samples from a triangular distribution described by shape parameters
defined by a lower limit, upper limit, and mode
}
\details{
Use vignette("fcmcmrr-class") for more information.
}
