% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{get_unconventional_adj_matrices_with_distributions}
\alias{get_unconventional_adj_matrices_with_distributions}
\title{get_unconventional_adj_matrices_with_distributions}
\usage{
get_unconventional_adj_matrices_with_distributions(
  unconventional_adj_matrix_list = list(matrix()),
  samples = 1000,
  include_zeroes = FALSE,
  show_progress = TRUE
)
}
\arguments{
\item{unconventional_adj_matrix_list}{A list of adjacency matrices with numeric and
either grey_number or triangular_number elements}

\item{samples}{The number of samples drawn from the representative pdf}

\item{include_zeroes}{TRUE/FALSE Whether or not to incorporate zeroes in
aggregation. If FALSE, zeroes are converted to NA values during aggregation and
returned to zero afterwards.}

\item{show_progress}{TRUE/FALSE Whether to display a progress bar at runtime}
}
\description{
Given a list of adjacency matrices which include either grey_numbers or
triangular_numbers, convert those objects to their corresponding
distributions representative of those values.
}
\details{
[ADD DETAILS HERE!!!!]

Use vignette("fcmconfr-class") for more information.
}
