% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{get_unconventional_value_as_distribution}
\alias{get_unconventional_value_as_distribution}
\title{get_unconventional_value_as_distribution}
\usage{
get_unconventional_value_as_distribution(
  value = numeric(),
  n_samples = 1000,
  value_class = ""
)
}
\arguments{
\item{value}{A grey_number or triangular_number object}

\item{n_samples}{The number of samples drawn from the representative pdf}

\item{value_class}{The class of the input value. Either grey_number or triangular_number. Can be left blank.}
}
\description{
Convert a grey/triangular number into a list of values sampled from a pdf of the
corresponding distribution (uniform for grey_numbers and triangular for triangular_numbers)
}
\details{
[ADD DETAILS HERE!!!!]

Use vignette("fcmconfr-class") for more information.
}
