% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{match_state_vector_df_shapes}
\alias{match_state_vector_df_shapes}
\title{match_state_vector_df_shapes}
\usage{
match_state_vector_df_shapes(baseline_state_vectors, scenario_state_vectors)
}
\arguments{
\item{baseline_state_vectors}{A state vectors dataframe for the baseline simulation}

\item{scenario_state_vectors}{A state vectors dataframe for the scenario simulation}
}
\description{
Given two data frames of state vectors, extend the one with the least number of rows
by repeating its final iteration value until the data frames are the same shape (i.e.
have the same number of rows)
}
\details{
Ensure that both input data frames are the same shape

Intended for developer use only to improve package readability.
}
