% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmr.R
\name{normalize_state_vector_with_optimized_lambda}
\alias{normalize_state_vector_with_optimized_lambda}
\title{normalize_state_vector_with_optimized_lambda}
\usage{
normalize_state_vector_with_optimized_lambda(
  raw_state = numeric(),
  squashed_state = numeric(),
  squashing = "sigmoid",
  optimized_lambda = numeric(),
  method = "none"
)
}
\arguments{
\item{raw_state}{The dot product of the state vector by a column vector of an adjacency matrix}

\item{squashed_state}{The output of a squashing function with the input raw state value}

\item{squashing}{A squashing function to apply. Must be one of the following: 'tanh', or 'sigmoid'.}

\item{optimized_lambda}{The optimized lambda calculated by optimize_fcm_lambda}

\item{method}{An algorithm of which to optimize lambda with. Must be one of the following: "koutsellis" or 'none'
if the user does not want to optimize lambda and use the user-defined lambda instead.

koutsellis: This algorithm was developed in Koutsellis et al. 2022 (https://doi.org/10.1109/IISA56318.2022.9904369).}
}
\description{
This calculates the normalized value of a state back into its originating
squashing function's domain rather than the one forced by the optimized lambda
}
\details{
This
}
