% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{old_salinization_ivfn_fcms}
\alias{old_salinization_ivfn_fcms}
\title{OLD! salinization_ivfn_fcms}
\format{
## `salinization_ivfn_fcms`
A list of conventional fcm adjacency matrices
}
\usage{
old_salinization_ivfn_fcms
}
\description{
An example fcm dataset collected by collecting stakeholder perceptions of
rising salinization in a socio-ecological system. To reduce file size and
protect stakeholder anonymity, the provided adjacency matrices contain a
subset of nodes and edges across the data used in the cited paper.
}
\details{
A buffer has been added to each edge weight so that they may be represented
as interval-value fuzzy numbers (IVFNs) instead of discrete numerical
values. The edge weight of the original, conventional fcms is contained
within the domain of all IVFN edge weights to allow for comparison.

Given in Rippy et al, (In Review)
}
\keyword{datasets}
