% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R
\name{plot.rtriangular_dist}
\alias{plot.rtriangular_dist}
\title{plot.rtriangular_dist}
\usage{
\method{plot}{rtriangular_dist}(x, ...)
}
\arguments{
\item{x}{an rtriangular_dist object}

\item{...}{additional inputs (leave empty)}
}
\description{
Plot rtriangular_dist distribution similar to how runif is plotted with the base plot function
}
