% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmconfr.R
\name{plot_ivfn_or_tfn_fcmconfr}
\alias{plot_ivfn_or_tfn_fcmconfr}
\title{Plot function for fcmconfr from ivfn or tfn FCM}
\usage{
plot_ivfn_or_tfn_fcmconfr(
  ivfn_or_tfn_fcmconfr_output,
  swap_axes = FALSE,
  aggregate_point_fill = "#fb0009",
  monte_carlo_col_fill = "#fcdbd0",
  monte_carlo_col_alpha = 0.6
)
}
\arguments{
\item{ivfn_or_tfn_fcmconfr_output}{Output directly from fcmconfr (for ivfn or tfn fcms as inputs)}

\item{swap_axes}{Switch x and y axes. Preferable when concepts have longer names}

\item{aggregate_point_fill}{Hex code value for the aggregate point fill}

\item{monte_carlo_col_fill}{Hex code value for the columns/bars representing monte carlo data}

\item{monte_carlo_col_alpha}{Between 0 and 1, the opacity of the monte_carlo_col_fill}
}
\description{
Plot function for fcmconfr from ivfn or tfn FCM
}
