% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fcms.R
\name{print.aggregate}
\alias{print.aggregate}
\title{Print method for aggregate}
\usage{
\method{print}{aggregate}(x, ...)
}
\arguments{
\item{x}{an aggregate object}

\item{...}{additional inputs}
}
\description{
Print method for aggregate
}
