% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmconfr.R
\name{print.fcmconfr}
\alias{print.fcmconfr}
\title{print.fcmconfr}
\usage{
\method{print}{fcmconfr}(x, ...)
}
\arguments{
\item{x}{an fgcm_simulation object}

\item{...}{additional inputs}
}
\description{
This improves the readability of the fcmconfr
}
\details{
Show the objects listed in the fcmconfr output $inference and $params, as well
as $bootstrap if present in output. Additionally, this prints descriptions/summaries
of objects within each sub-list like inference_opts, bootstrap_input_opts, etc.

Use vignette("fcm-class") for more information.
}
