% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.fcmconfr_gui.R
\name{print.fcmconfr_gui_input}
\alias{print.fcmconfr_gui_input}
\title{Print fcmconfr_gui_input}
\usage{
\method{print}{fcmconfr_gui_input}(x, ...)
}
\arguments{
\item{x}{an fcmconfr object}

\item{...}{additional inputs}
}
\value{
A pasteable output to the console of a call to fcmconfr with the
selected variables
}
\description{
This function prints a clean, copy-and-paste-able call to \code{\link{fcmconfr}}
with the inputs selected by the user in the shiny app.
}
\examples{
NULL
}
