% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcmr.R
\name{print.fgcmr_simulation}
\alias{print.fgcmr_simulation}
\title{print.fgcmr_simulation}
\usage{
\method{print}{fgcmr_simulation}(x, ...)
}
\arguments{
\item{x}{an fgcmr_simulation object}

\item{...}{additional inputs}
}
\description{
This improves the readability of the simulate_fgcmr output
}
\details{
Show the first two iterations of the simulation, followed by a gap, and then
the final state vector in an organized data frame. Additionally, show the
activation, squashing, lambda, and algorithm inputs as well as the total
number of iterations and goal minimum error.

Use vignette("fgcmr-class") for more information.
}
