% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgcmr.R
\name{print.grey_number}
\alias{print.grey_number}
\title{print.grey_number}
\usage{
\method{print}{grey_number}(x)
}
\arguments{
\item{lower}{lower limit of a Grey Number set (the lower value must be less than or equal to the upper value)}

\item{upper}{upper limit of a Grey Number set (the upper value must be greater or equal to the lower value)}
}
\description{
This improves the readability of the output
}
\details{
Grey Numbers represent intervals within a possibility space (typically
[0, 1] or [-1, 1]) that contains the true value of an object.

The grey_number class does not perform any operations on its input, rather
it checks whether the input follows the defining criteria of grey numbers

grey_numbers are constructed using vctrs::new_vctr instead of the typical call to
structure() because ...

Use vignette("fgcmr-class") for more information.
}
\examples{
grey_number(lower = 0, upper = 1)
}
