% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_and_simulate_fcm.R
\name{print.infer_conventional_fcm}
\alias{print.infer_conventional_fcm}
\title{Print method for infer_conventional_fcm objects}
\usage{
\method{print}{infer_conventional_fcm}(x, ...)
}
\arguments{
\item{x}{an infer_conventional_fcm object}

\item{...}{additional inputs}
}
\description{
Print method for infer_conventional_fcm objects
}
