% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R,
%   R/[DEPRECATED]interval_value_fuzzy_numbers.R
\name{print.ivfn}
\alias{print.ivfn}
\title{print.ivfn}
\usage{
\method{print}{ivfn}(x, ...)

\method{print}{ivfn}(x, ...)
}
\arguments{
\item{x}{a ivfn object}

\item{...}{additional inputs}
}
\description{
This improves the readability of the output

This improves the readability of the output
}
\details{
Grey Numbers represent intervals within a possibility space (typically
[0, 1] or [-1, 1]) that contains the true value of an object.

The ivfn class does not perform any operations on its input, rather
it checks whether the input follows the defining criteria of grey numbers

ivfns are constructed using vctrs::new_vctr instead of the typical call to
structure() because ...

Use vignette("fgcm-class") for more information.

Grey Numbers represent intervals within a possibility space (typically
[0, 1] or [-1, 1]) that contains the true value of an object.

The ivfn class does not perform any operations on its input, rather
it checks whether the input follows the defining criteria of grey numbers

ivfns are constructed using vctrs::new_vctr instead of the typical call to
structure() because ...

Use vignette("fgcm-class") for more information.
}
\examples{
ivfn(lower = 0, upper = 1)
ivfn(lower = 0, upper = 1)
}
