% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_value_fuzzy_numbers.R
\name{print.ivfn_simulation}
\alias{print.ivfn_simulation}
\title{print.fgcm_simulation}
\usage{
\method{print}{ivfn_simulation}(x, ...)
}
\arguments{
\item{x}{an fgcm_simulation object}

\item{...}{additional inputs}
}
\description{
This improves the readability of the infer_fgcm_with_pulse output
}
\details{
Show the first two iterations of the simulation, followed by a gap, and then
the final state vector in an organized data frame. Additionally, show the
activation, squashing, lambda, and algorithm inputs as well as the total
number of iterations and goal minimum error.

Use vignette("fgcm-class") for more information.
}
