% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_triangular_cognitive_maps.R
\name{print.triangular_number}
\alias{print.triangular_number}
\title{print.triangular_number}
\usage{
\method{print}{triangular_number}(x, ...)
}
\arguments{
\item{x}{a triangular_number object}

\item{...}{additional inputs}
}
\description{
This improves the readability of the output
}
\details{
triangular Numbers represent intervals within a possibility space (typically
[0, 1] or [-1, 1]) that contains the true value of an object.

The triangular_number class does not perform any operations on its input, rather
it checks whether the input follows the defining criteria of triangular numbers

triangular_numbers are constructed using vctrs::new_vctr instead of the typical call to
structure() because ...

Use vignette("ftcm-class") for more information.
}
\examples{
triangular_number(lower = 0, upper = 1)
}
