% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R
\name{rtri}
\alias{rtri}
\title{The Triangular Distribution}
\usage{
rtri(n, lower = 0, mode = 0.5, upper = 1)
}
\arguments{
\item{n}{number of samples to draw from the triangular distribution}

\item{lower}{lower limit or minimum of the sample space}

\item{mode}{peak of the sample space}

\item{upper}{upper limit or maximum of the sample space}
}
\value{
a set c() of values representing a triangular distribution
}
\description{
This function generates a triangular distribution on the interval from lower
to upper bounds with a a given mode,
}
\details{
Additional details...
}
