% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R
\name{rtriangular_dist}
\alias{rtriangular_dist}
\title{The Triangular Distribution}
\usage{
rtriangular_dist(
  n = integer(),
  lower = double(),
  mode = double(),
  upper = double()
)
}
\arguments{
\item{n}{number of samples to draw from the triangular distribution}

\item{lower}{lower limit or minimum of the sample space}

\item{mode}{peak of the sample space}

\item{upper}{upper limit or maximum of the sample space}
}
\value{
a set c() of values representing a triangular distribution
}
\description{
This function generates a triangular distribution on the interval from lower
to upper bounds with a a given mode,
}
\details{
Additional details...
}
\examples{
rtriangular_dist(100, -1, 0, 1)

x <- rtriangular_dist(n = 1000, lower = -0.5, mode = 0, upper = 0.5)
hist(x)

x <- rtriangular_dist(100, 0, 0, 0.5)
hist(x)
}
