% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salinization_ses_fcms}
\alias{salinization_ses_fcms}
\title{salinization_ses_fcms}
\format{
## `salinization_ses_fcms`
A list of fcm adjacency matrices
}
\usage{
salinization_ses_fcms
}
\description{
An example fcm dataset collected by collecting stakeholder perceptions of
rising salinization in a socio-ecological system.
}
\details{
Given in Rippy et al, (2024)
}
\keyword{datasets}
