% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salinization_tfn_fcms}
\alias{salinization_tfn_fcms}
\title{salinization_tfn_fcms}
\format{
## `salinization_tfn_fcms`
A list of conventional fcm adjacency matrices
}
\usage{
salinization_tfn_fcms
}
\description{
An example fcm dataset collected by collecting stakeholder perceptions of
rising salinization in a socio-ecological system. To reduce file size and
protect stakeholder anonymity, the provided adjacency matrices contain a
subset of nodes and edges across the data used in the cited paper.
}
\details{
A buffer has been added to each edge weight so that they may be represented
as triangular-value fuzzy numbers (TFNs) instead of discrete numerical
values. The edge weight of the original, conventional fcms is contained
as the mode value of all TFN edge weights to allow for comparison.

Given in Rippy et al, (In Review)
}
\keyword{datasets}
