% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_fgcm}
\alias{simulate_fgcm}
\title{simulate_fgcm}
\usage{
simulate_fgcm(
  grey_adj_matrix = matrix(),
  initial_state_vector = vector(),
  iter = 10,
  f_squashing = "sigmoid"
)
}
\arguments{
\item{grey_adj_matrix}{An adjacency matrix representing an fgcm, made up of
grey_number type objects}

\item{initial_state_vector}{A list of initial state values for an fgcm}

\item{iter}{Number of iterations to perform}

\item{f_squashing}{Squashing function}
}
\description{
This simulates how an fgcm reacts to an initial state vector
}
\details{
X

#' Use vignette("fgcmr-class") for more information.
}
\examples{
NULL
}
