% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmcmcmr.R
\name{simulate_fmcmcmr_models}
\alias{simulate_fmcmcmr_models}
\title{simulate_fmcmcmr_models}
\usage{
simulate_fmcmcmr_models(
  simulated_adj_matrices = list(matrix()),
  initial_state_vector = c(),
  activation = "modified-kosko",
  squashing = "sigmoid",
  lambda = 1,
  max_iter = 10,
  min_error = 1e-05,
  lambda_optimization = "none",
  IDs = c(),
  parallel = TRUE
)
}
\arguments{
\item{simulated_adj_matrices}{A list of adjecency matrices generated from simulation using build_fmcmcmr_models.}

\item{initial_state_vector}{A list state values at the start of an fcm simulation}

\item{activation}{The activation function to be applied. Must be one of the following:
'kosko', 'modified-kosko', or 'papageorgiou'.}

\item{squashing}{A squashing function to apply. Must be one of the following:
'bivalent', 'saturation', 'trivalent', 'tanh', or 'sigmoid'.}

\item{lambda}{A numeric value that defines the steepness of the slope of the
squashing function when tanh or sigmoid are applied}

\item{max_iter}{The maximum number of iterations to run if the minimum error value is not achieved}

\item{min_error}{The lowest error (sum of the absolute value of the current state
vector minus the previous state vector) at which no more iterations are necessary
and the simulation will stop}

\item{lambda_optimization}{A lambda optimization procedure to apply. Must be one
of the following: 'none' or 'koutsellis'}

\item{IDs}{A list of names for each node (must have n items). If empty, will use
column names of adjacancy matrix (if given).}

\item{parallel}{TRUE/FALSE Whether to utilize parallel processing}
}
\description{
This calculates a sequence of iterations of a simulation over every item in
a list of fmcmcmr objects given an initial state vector along with the
activation, squashing, and lambda parameters.
Additional variables may be defined to control simulation length,
column names, and lambda optimization.
}
\details{
[ADD DETAILS HERE!!!]

Use vignette("fmcmcmr-class") for more information.
}
