% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_fcms.R
\name{squash}
\alias{squash}
\title{squash}
\usage{
squash(value = numeric(), squashing = "sigmoid", lambda = 1)
}
\arguments{
\item{value}{A numeric value to 'squash'}

\item{squashing}{A squashing function to apply. Must be one of the following: 'bivalent', 'saturation', 'trivalent', 'tanh', or 'sigmoid'}

\item{lambda}{A numeric value that defines the steepness of the slope of the squashing function when tanh or sigmoid are applied}
}
\description{
Calculate squashing function output of an input value and lambda values
}
\details{
This function calculates the 'squashed' value of a state based upon five
available squashing functions typical in the literature (as identified in
Gonzales et al. 2018 - https://doi.org/10.1142/S0218213018600102)
}
