% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R
\name{subtract_ivfn}
\alias{subtract_ivfn}
\title{IVFN Subtraction (Interval Calculus)}
\usage{
subtract_ivfn(ivfn_1 = ivfn(), ivfn_2 = ivfn())
}
\arguments{
\item{ivfn_1}{An interval-value fuzzy number (ivfn) object}

\item{ivfn_2}{An interval-value fuzzy number (ivfn) object}
}
\value{
An IVFN object representing the subtraction of ivfn_2 from ivfn_1
}
\description{
This subtracts one Interval-Value Fuzzy Number (IVFN) from another via
interval calculus. An IVFN represented by the interval X:

\deqn{X\  =( x_{1}, x_{3}) \ =\ [ x_{1} ;x_{2}] =\{x\in \mathbb{R} \ |\ x_{1} \ \leq x\ \leq x_{2}\}}{ascii}

where \eqn{x_1}{ascii} and \eqn{x_2}{ascii} represent the lower and upper
bounds, respectively.

The IVFN X may have another IVFN Y subtracted from it via:

\deqn{X\ -\ Y\ =\ [ x_{1} \ -\ y_{2} ;\ x_{2} \ -\ y_{1}]}{ascii}
}
\details{
It is not required for one IVFN to be "greater than" the other.

This difference may also be estimated by translating the IVFN's into
their corresponding distributions (e.g. ivfn(-1, 1) = runif(n, -1, 1)),
subtracting one distribution from the other, and estimating the minimum
and maximum values of the difference distribution.
}
\examples{
subtract_ivfn(ivfn(0.5, 0.8), ivfn(0.2, 0.5))
subtract_ivfn(ivfn(-0.5, 0.3), ivfn(0.4, 0.6))
subtract_ivfn(ivfn(-1, 1), ivfn(-0.5, 0.5))
}
\references{
\insertRef{mooreIntervalAnalysisFuzzy2003}{fcmconfr}

\insertRef{dimuroIntervalFuzzyNumbers2011}{fcmconfr}

\insertRef{mooreIntervalAnalysis1966}{fcmconfr}
}
\seealso{
Other interval-valued-fuzzy-numbers: 
\code{\link{c.ivfn}()},
\code{\link{ivfn}()},
\code{\link{make_adj_matrix_w_ivfns}()},
\code{\link{print.ivfn}()}
}
\concept{interval-valued-fuzzy-numbers}
