% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R
\name{subtract_tfn}
\alias{subtract_tfn}
\title{TFN Subtraction (Interval Calculus)}
\usage{
subtract_tfn(tfn_1 = tfn(), tfn_2 = tfn())
}
\arguments{
\item{tfn_1}{A triangular fuzzy number (tfn) object}

\item{tfn_2}{A triangular fuzzy number (tfn) object}
}
\value{
An TFN object representing the subtraction of tfn_2 from tfn_1
}
\description{
This subtracts one Triangular Fuzzy Number (TFN) from another via
interval calculus. A TFN represented by the fuzzy set (triangular distribution):

\deqn{X\  =( x_{1} ,x_{2} ,x_{3}) \ =\ \left\{\begin{matrix}
0 & for\  & x< x_{1}\\
\frac{2( x\ -\ x_{1})}{( x_{3} -x_{1})( x_{2} -x_{1})} & for\  & x_{1} \leq x< x_{3} \ \\
\frac{2}{x_{3} -x_{1}} & for & x\ =\ x_{2}\\
\frac{2( x_{3} -x)}{( x_{3} -x_{1})( x_{3} -x_{2})} & for\  & x_{2} < x\leq x_{3}\\
0 & for & x >x_{3}
\end{matrix}\right.}{ascii}

where \eqn{x_1}{ascii} and \eqn{x_3}{ascii} are the lower and upper bounds, respectively,
and \eqn{x_2}{ascii} is the mode.

The TFN X may have another TFN Y subtracted from it via:

\deqn{X\ -\ Y\ =\ ( x_{1} -y_{3} ,\ x_{2} -y_{2} ,\ x_{3} -y_{1})}{ascii}
}
\details{
It is not required for one IVFN to be "greater than" the other.

This difference may also be estimated by translating the TFN's into
their corresponding distributions
(e.g. tfn(-1, 0, 1) = EnvStats::rtri(n, min = -1, max = 1, mode = 0)),
subtracting one distribution from the other, and estimating the minimum,
mode, and maximum values of the difference distribution.
}
\examples{
subtract_tfn(tfn(0.5, 0.6, 0.8), tfn(0.2, 0.3, 0.5))
subtract_tfn(tfn(-0.5, -0.2, 0.3), tfn(0.4, 0.5, 0.6))
subtract_tfn(tfn(-1, 0, 1), tfn(-0.5, 0, 0.5))
}
\references{
\insertRef{chakravertyFuzzyNumbers2019}{fcmconfr}

\insertRef{hanssAppliedFuzzyArithmetic2005}{fcmconfr}

\insertRef{trillasFuzzyArithmetic2015}{fcmconfr}
}
\seealso{
Other triangular-fuzzy-numbers: 
\code{\link{make_adj_matrix_w_tfns}()},
\code{\link{print.tfn}()},
\code{\link{tfn}()}
}
\concept{triangular-fuzzy-numbers}
