% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R,
%   R/[DEPRECATED]triangular_fuzzy_numbers.R
\name{tfn}
\alias{tfn}
\title{tfn S3 class}
\usage{
tfn(lower = double(), mode = double(), upper = double())

tfn(lower = double(), mode = double(), upper = double())
}
\arguments{
\item{lower}{lower limit of a Triangular Number set (the lower value must be less than or equal to the upper value)}

\item{mode}{the most likely value of a Triangular Number set}

\item{upper}{upper limit of a Triangular Number set (the upper value must be greater or equal to the lower value)}
}
\description{
This class is an organization scheme for Triangular Numbers (See
____, XXXX for example).

This class is an organization scheme for Triangular Numbers (See
____, XXXX for example).
}
\details{
Triangular Numbers represent intervals within a possibility space (typically
[0, 1] or [-1, 1]) that contains the true value of an object.

The tfn class does not perform any operations on its input, rather
it checks whether the input follows the defining criteria of triangular numbers

Use vignette("fcm_w_fcm_w_tfn-class") for more information.

Triangular Numbers represent intervals within a possibility space (typically
[0, 1] or [-1, 1]) that contains the true value of an object.

The tfn class does not perform any operations on its input, rather
it checks whether the input follows the defining criteria of triangular numbers

Use vignette("fcm_w_fcm_w_tfn-class") for more information.
}
\examples{
tfn(lower = 0, mode = 0.5, upper = 1)
tfn(lower = 0, mode = 0.5, upper = 1)
}
