% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcmconfr.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom plot.fcmconfr Theme}
\usage{
theme_custom(...)
}
\arguments{
\item{...}{Additional Inputs}
}
\description{
Theme-ing for plot.fcmconfr to improve readability in
\code{\link{plot.fcmconfr}} function definition
}
