% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_triangular_cognitive_maps.R
\name{triangular_number}
\alias{triangular_number}
\title{triangular_number S3 class}
\usage{
triangular_number(lower = double(), mode = double(), upper = double())
}
\arguments{
\item{lower}{lower limit of a Triangular Number set (the lower value must be less than or equal to the upper value)}

\item{mode}{the most likely value of a Triangular Number set}

\item{upper}{upper limit of a Triangular Number set (the upper value must be greater or equal to the lower value)}
}
\description{
This class is an organization scheme for Triangular Numbers (See
____, XXXX for example).
}
\details{
Triangular Numbers represent intervals within a possibility space (typically
[0, 1] or [-1, 1]) that contains the true value of an object.

The triangular_number class does not perform any operations on its input, rather
it checks whether the input follows the defining criteria of triangular numbers

Use vignette("ftcm-class") for more information.
}
\examples{
triangular_number(lower = 0, mode = 0.5, upper = 1)
}
