% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vat_fgcm_examples}
\alias{vat_fgcm_examples}
\title{Vat FGCM Examples}
\format{
## `vat_fgcm_examples`
A list of fgcm adjacency matrices and initial state vectors
\describe{
  \item{fcm_as_fgcm}{An fgcm where every value is a white number to allow
  users to compare results with ordinary fcm}
  \item{fgcm}{An fgcm with grey values incorporated in its adjacency matrix}
  ...
}
}
\source{
<https://doi.org/10.1007/s10489-013-0511-z>
}
\usage{
vat_fgcm_examples
}
\description{
An example fgcm representing a multi-tank vat system. Depicted in
Salmeron & Papageorgiou, 2014 - https://doi.org/10.1007/s10489-013-0511-z
}
\keyword{datasets}
