% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fcms.R
\name{aggregate_fcms}
\alias{aggregate_fcms}
\title{aggregate_fcms}
\usage{
aggregate_fcms(
  adj_matrices = list(matrix()),
  aggregation_function = c("mean", "median"),
  include_zeroes = FALSE
)
}
\arguments{
\item{adj_matrices}{A list type object of fcms. Must have a length greater
than 1.}

\item{aggregation_function}{"mean" or "median"}

\item{include_zeroes}{TRUE/FALSE Whether to include zeroes in the mean/median
calculations. (i.e. if edges not included in a map should count as a zero-weighted
edge or not at all)}
}
\description{
aggregate_fcms
}
