% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_fcms.R
\name{aggregate_fcms_w_ivfns}
\alias{aggregate_fcms_w_ivfns}
\title{Aggregate (IVFN) FCMs}
\usage{
aggregate_fcms_w_ivfns(
  adj_matrices = list(matrix()),
  aggregation_function = c("mean", "median"),
  include_zeroes = TRUE
)
}
\arguments{
\item{adj_matrices}{A list of IVFN FCM adj. matrix objects}

\item{aggregation_function}{Calculate aggregate edge weights as either the
"mean" or "median" of the input edge weights across inputs}

\item{include_zeroes}{TRUE/FALSE Whether to include zeroes in the mean/median
calculations. (i.e. if edges not included in a map should count as a zero-weighted
edge or not at all)}
}
\value{
An aggregate adj. matrix (of class 'aggregate') with edges represented
as IVFN data types
}
\description{
Generate an aggregate adj. matrix from a list of (IVFN) adj. matrices.
FCM aggregation works by calculating the mean/median edge weight for all edges
across the input adj. matrices (i.e. the mean/median of the edge weight
connecting A -> B across all maps, the mean/median of the edge weight
connecting B -> C across all maps, and so on). The user may dictate whether
to incorporate 0-valued edge weights in the mean/median calculations.
}
\details{
All input adj. matrices must represent IVFN FCMs
All input adj. matrices must have the same dimensions and concept names to
generate an aggregate.
}
\examples{
lower_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
upper_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
adj_matrix_1 <- make_adj_matrix_w_ivfns(lower_adj_matrix_1, upper_adj_matrix_1)
lower_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
upper_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(1, 0)
)
adj_matrix_2 <- make_adj_matrix_w_ivfns(lower_adj_matrix_2, upper_adj_matrix_2)
lower_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.2, 0)
)
upper_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
adj_matrix_3 <- make_adj_matrix_w_ivfns(lower_adj_matrix_3, upper_adj_matrix_3)
lower_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.0, 0)
)
upper_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
adj_matrix_4 <- make_adj_matrix_w_ivfns(lower_adj_matrix_4, upper_adj_matrix_4)
fcms_w_ivfns <- list(adj_matrix_1, adj_matrix_2, adj_matrix_3, adj_matrix_4)

# Note that Row (1), Column(2) of adj_matrix 3 has a false-zero lower
# edge weight. See how the aggregate incorporates those values even though
# include_zeroes = FALSE
aggregate_fcms_w_ivfns(fcms_w_ivfns, "mean", include_zeroes = FALSE)
aggregate_fcms_w_ivfns(fcms_w_ivfns, "mean", include_zeroes = TRUE)
}
\seealso{
Other aggregate_fcms: 
\code{\link{aggregate_conventional_fcms}()},
\code{\link{aggregate_fcms}()},
\code{\link{aggregate_fcms_w_tfns}()},
\code{\link{print.aggregate}()}
}
\concept{aggregate_fcms}
