% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcmconfr.R
\name{autoplot.fcmconfr}
\alias{autoplot.fcmconfr}
\title{Autoplot fcmconfr}
\usage{
\method{autoplot}{fcmconfr}(
  fcmconfr_object,
  coord_flip = TRUE,
  aggregate_fill = "#f7e3dd",
  aggregate_color = "#E1D1CB",
  aggregate_alpha = 0.7,
  mc_fill = "#f7e3dd",
  ...
)
}
\description{
Autoplot fcmconfr
}
