% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{build_monte_carlo_fcms_from_conventional_adj_matrices}
\alias{build_monte_carlo_fcms_from_conventional_adj_matrices}
\title{Build Monte Carlo (Conventional) FCMs}
\usage{
build_monte_carlo_fcms_from_conventional_adj_matrices(
  adj_matrix_list = list(Matrix::sparseMatrix()),
  N_samples = integer(),
  include_zeroes = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{adj_matrix_list}{A list of n x n adjacencey matrices representing fcms}

\item{N_samples}{The number of samples to draw with the selected sampling method. Also,
the number of sampled models to generate}

\item{include_zeroes}{TRUE/FALSE Whether to incorporate zeroes as intentionally-defined
edge weights or ignore them in aggregation}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}
}
\value{
A list of empirical (Conventional) FCM adj. matrices generated via monte carlo methods
}
\description{
This function generates n fcm models whose edge weights are sampled from either
the defined edge values in a set of adjacency matrices derived from the sets
of edge values, and stores them as a list of adjacency matrices.
}
\details{
The show_progress and parallel inputs change the functions called, but do NOT
change the output! These are allowed to be toggled on/off to increase user
control at runtime.
}
\examples{
ex_adj_matrix_1 <- data.frame(
  "A" = c(0, 0, 0, 0),
  "B" = c(0, 0, 0, 1),
  "C" = c(0, 1, 0, 0),
  "D" = c(0, 0, 1, 0)
)
ex_adj_matrix_2 <- data.frame(
  "A" = c(0, 0, 0, 0),
  "B" = c(0.25, 0, 0, 0.25),
  "C" = c(0, 0.25, 0, 0),
  "D" = c(0, 0, 0.25, 0)
)
ex_adj_matrix_3 <- data.frame(
  "A" = c(0, 0, 0, 0),
  "B" = c(0.75, 0, 0, 0.75),
  "C" = c(0, 0.75, 0, 0),
  "D" = c(0, 0, 0.75, 0)
)
ex_adj_matrix_4 <- data.frame(
  "A" = c(0, 0, 0, 0),
  "B" = c(0.5, 0, 0, 0.5),
  "C" = c(0, 0.5, 0, 0),
  "D" = c(0, 0, 0.5, 0)
)
ex_fcms <- list(
  ex_adj_matrix_1, ex_adj_matrix_2, ex_adj_matrix_3, ex_adj_matrix_4
)

mc_conventional_fcms <- build_monte_carlo_fcms_from_conventional_adj_matrices(
  ex_fcms, N_samples = 1000, include_zeroes = FALSE, show_progress = TRUE
)
}
\seealso{
Other monte-carlo-model-generation-and-simulation: 
\code{\link{build_monte_carlo_fcms}()},
\code{\link{build_monte_carlo_fcms_from_fuzzy_set_adj_matrices}()},
\code{\link{get_mc_simulations_inference_CIs_w_bootstrap}()},
\code{\link{infer_monte_carlo_fcm_set}()},
\code{\link{monte_carlo_bootstrap_checks}()}
}
\concept{monte-carlo-model-generation-and-simulation}
