% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_and_simulate_fcm.R
\name{convert_element_to_ivfn_or_tfn_if_numeric}
\alias{convert_element_to_ivfn_or_tfn_if_numeric}
\title{Convert Value to IVFN or TFN if Value is Numeric}
\usage{
convert_element_to_ivfn_or_tfn_if_numeric(
  element,
  desired_class = c("ivfn", "tfn")
)
}
\arguments{
\item{element}{An element in a matrix}

\item{desired_class}{Transform the element into an 'ivfn' or 'tfn'}
}
\value{
An IVFN or TFN representation of a crisp, numeric value
}
\description{
This checks whether the input element is an ordinary number or a triangular number.
If it is an ivfn or tfn, it returns the input, but if it is a numeric type
object (ordinary number), it will convert that number into an ivfn or tfn
}
\examples{
convert_element_to_ivfn_or_tfn_if_numeric(0.6, "ivfn")
convert_element_to_ivfn_or_tfn_if_numeric(0.7, "tfn")
}
