% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{convert_fuzzy_elements_in_matrix_to_distributions}
\alias{convert_fuzzy_elements_in_matrix_to_distributions}
\title{}
\usage{
convert_fuzzy_elements_in_matrix_to_distributions(
  fuzzy_matrix = data.frame(),
  fuzzy_element_class = c("fgcm", "fcm_w_tfn"),
  N_samples = integer()
)
}
\arguments{
\item{fuzzy_matrix}{A matrix that can contain fuzzy sets as elements}

\item{fuzzy_element_class}{"fgcm" or "fcm_w_tfn" - the class of elements in the fuzzy_matrix}

\item{N_samples}{The number of samples to draw from the corresponding distribution}
}
\description{
Given a list of adjacency matrices which include either grey_numbers or
tfns, convert those objects to their corresponding
distributions representative of those values.
}
\details{

}
