% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_and_simulate_fcm.R
\name{convert_fuzzy_set_elements_in_matrix_to_distributions}
\alias{convert_fuzzy_set_elements_in_matrix_to_distributions}
\title{Convert IVFN or TFN Elements in Adj. Matrix to Distributions (i.e. sets)}
\usage{
convert_fuzzy_set_elements_in_matrix_to_distributions(
  fuzzy_set_matrix = matrix(),
  object_class = c("ivfn", "tfn"),
  N_samples = integer()
)
}
\arguments{
\item{fuzzy_set_matrix}{A matrix that contains fuzzy sets as elements}

\item{object_class}{Values are represented either as ivfns or tfns. Options: 'ivfn' or 'tfn'}

\item{N_samples}{The number of samples to draw from the corresponding distribution}
}
\value{
An adj. matrix of IVFNs or TFNs represented as lists (sets) of their
representative distributions
}
\description{
Given a list of adjacency matrices which include either ivfns or
tfns, convert those objects to their corresponding
distributions representative of those values.
}
\details{
This function assists with subtracting the baseline from the scenario
simulation when calling infer_fcm with IVFN-FCMs or TFN-FCMs.
}
