% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_and_simulate_fcm.R
\name{convert_fuzzy_set_elements_in_matrix_to_distributions}
\alias{convert_fuzzy_set_elements_in_matrix_to_distributions}
\title{convert_fuzzy_set_elements_in_matrix_to_distributions}
\usage{
convert_fuzzy_set_elements_in_matrix_to_distributions(
  fuzzy_set_matrix = matrix(),
  object_class = c("ivfn", "tfn"),
  N_samples = integer()
)
}
\arguments{
\item{fuzzy_set_matrix}{A matrix that contains fuzzy sets as elements}

\item{object_class}{Values are represented either as ivfns or tfns. Options: 'ivfn' or 'tfn'}

\item{N_samples}{The number of samples to draw from the corresponding distribution}
}
\description{
Given a list of adjacency matrices which include either grey_numbers or
tfns, convert those objects to their corresponding
distributions representative of those values.
}
\details{
[ADD DETAILS HERE!!!!]

Use vignette("fcmconfr-class") for more information.
}
