% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_fuzzy_numbers.R
\name{defuzz_tfn}
\alias{defuzz_tfn}
\title{defuzz_tfn}
\usage{
defuzz_tfn(tfn_object = tfn())
}
\arguments{
\item{tfn_object}{a tfn object (generated with tfn())}
}
\description{
Translate a triangular fuzzy number (tfn) into a "crisp" value by calculating
the average of the lower, mode, and upper values.
}
