% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmconfr.R
\name{fcmconfr}
\alias{fcmconfr}
\title{fcmconfr}
\usage{
fcmconfr(
  adj_matrices = list(matrix()),
  aggregation_function = c("mean", "median"),
  monte_carlo_sampling_draws = 1000,
  initial_state_vector = c(),
  clamping_vector = c(),
  activation = c("kosko", "modified-kosko", "rescale"),
  squashing = c("sigmoid", "tanh"),
  lambda = 1,
  max_iter = 100,
  min_error = 1e-05,
  fuzzy_set_samples = 1000,
  inference_estimation_CI = 0.95,
  inference_estimation_bootstrap_reps = 5000,
  inference_estimation_bootstrap_draws_per_rep = 5000,
  show_progress = TRUE,
  parallel = TRUE,
  n_cores = integer(),
  include_zero_weighted_edges_in_aggregation_and_mc_sampling = FALSE,
  include_monte_carlo_FCM_simulations_in_output = TRUE,
  estimate_mc_inference_CI_w_bootstrap = TRUE
)
}
\arguments{
\item{adj_matrices}{A list of adjacency matrices (n x n) representing FCMs. This
can also be an individual adjacency matrix.Adj. Matrices can be conventional FCMs,
FCMs with edge weights as Interval Value Fuzzy Numbers (IVFNs) or FCMs with edge
weights as Triangular Fuzzy Numbers (TFNs)}

\item{aggregation_function}{Aggregate the adj. matrices into a single FCM by taking
either the mean or median of the edge weights for edges included in multiple maps}

\item{monte_carlo_sampling_draws}{The number of FCMs to generate via monte carlo
sampling from the input adj. matrices}

\item{initial_state_vector}{A list state values at the start of an fcm simulation}

\item{clamping_vector}{A list of values representing specific actions taken to
control the behavior of an FCM. Specifically, non-zero values defined in this vector
will remain constant throughout the entire simulation as if they were "clamped" at those values.}

\item{activation}{The activation function to be applied. Must be one of the following:
'kosko', 'modified-kosko', or 'papageorgiou'.}

\item{squashing}{A squashing function to apply. Must be one of the following:
'bivalent', 'saturation', 'trivalent', 'tanh', or 'sigmoid'.}

\item{lambda}{A numeric value that defines the steepness of the slope of the
squashing function when tanh or sigmoid are applied}

\item{max_iter}{The maximum number of iterations to run if the minimum error value is not achieved}

\item{min_error}{The lowest error (sum of the absolute value of the current state
vector minus the previous state vector) at which no more iterations are necessary
and the simulation will stop}

\item{fuzzy_set_samples}{The size (n) of the distributions represented by IVFNs or TFNs (only
used when IVFNs or TFNs in adj_matrices input)}

\item{inference_estimation_CI}{The confidence interval to estimate for the inferences
of each concept across all monte carlo FCMs (via bootstrap)}

\item{inference_estimation_bootstrap_reps}{The number of bootstraps to perform in
estimating the confidence interval for the inferences of each concept across all monte
carlo FCMs}

\item{inference_estimation_bootstrap_draws_per_rep}{The number of draws to sample
(with replacement) from the estimates for the inferences of each concept across
all monte carlo FCMs per bootstrap}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}

\item{parallel}{TRUE/FALSE Whether to utilize parallel processing}

\item{n_cores}{Number of cores to use in parallel processing. If no input given,
will use all available cores in the machine.}

\item{include_zero_weighted_edges_in_aggregation_and_mc_sampling}{TRUE/FALSE
Whether to incorporate zeroes as intentionally-defined edge weights or ignore
them when aggregating adj. matrices and sampling for monte carlo FCMs}

\item{include_monte_carlo_FCM_simulations_in_output}{TRUE/FALSE
Whether to include simulations of monte carlo FCMs. Switch to FALSE if concerned
about the size of the output of fcmconfr (simulations are necessary and will run regardless)}

\item{estimate_mc_inference_CI_w_bootstrap}{TRUE/FALSE
Whether to estimate confidence intervals for the inferences of each concept across
all monte carlo FCMs via bootstrap. Switch to FALSE if concerned about runtime
and/or want to perform bootstrapping manually.}
}
\description{
[ADD DETAILS HERE!!!]
}
\details{
[ADD DETAILS HERE!!!]

Use vignette("fcmconfr") for more information.
}
