% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmcmcmr.R
\name{fmcmcmr}
\alias{fmcmcmr}
\title{fmcmcmr (fuzzy monte carlo markov chain cognitive map) S3 class}
\usage{
fmcmcmr(adj_matrix = matrix(), IDs = c(), distribution = "uniform", ...)
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents the edge weights
of an FCM}

\item{IDs}{A list of names for each node (must have n items)}

\item{distribution}{A statistical distribution to draw random samples from.
Must be one of the following: 'uniform', 'gaussian', 'beta', or 'triangular'}

\item{...}{Additional adj_matrix objects whose weights describe shape parameters
of the chosen distribution.
IF distribution = "uniform", must include: lower_adj_matrix and upper_adj_matrix objects
IF distribution = "gaussian", must include: sd_adj_matrix
IF distribution = "beta", must include: sd_adj_matrix
IF distribution = "triangular", must include: lower_adj_matrix, upper_adj_matrix, and mode_adj_matrix objects
 Note: if no mode_adj_matrix given, will assume its values are the adj_matrix values.}
}
\description{
This class is an organization scheme for fuzzy cognitive maps that model
uncertainty using monte carlo markov chain simulations (See
Koasidis et al. 2022 - https://dx.doi.org/10.2139/ssrn.4233987 or
Koutsellis et al. 2023 - https://doi.org/10.1016/j.softx.2023.101513 for
an application in python.

###It stores the nodes of an FCM and its corresponding adjacency matrix
and edgelist.
}
\details{
fmcmcmr stores fmcmcm data in forms useful to data manipulation, particular
regarding pairing with popular network analysis libraries like igraph
or visNetwork.

Use vignette("fmcmcmr-class") for more information.
}
