% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcmconfr.R
\name{get_concepts_to_plot}
\alias{get_concepts_to_plot}
\title{Get Concepts in fcmconfr Object to Include in Plot}
\usage{
get_concepts_to_plot(fcmconfr_object, filter_limit = 0.01)
}
\description{
Get Concepts in fcmconfr Object to Include in Plot
}
