% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_edgelist_from_adj_matrix}
\alias{get_edgelist_from_adj_matrix}
\title{Convert an Adjacency Matrix to an Edgelist}
\usage{
get_edgelist_from_adj_matrix(adj_matrix = matrix())
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents an FCM}
}
\value{
An edgelist with the following columns: source, target, weight
}
\description{
Converts an adjacency matrix into an edgelist
}
\details{
The input adjacency matrix must be a square n x n matrix. It can be either
a matrix, data.frame, tibble, or data.table type object.

If the input matrix has named columns, those names will be used as concepts
in the edgelist. Otherwise, generic node IDs will be used (C1, C2, ... Cn)
}
\examples{
adj_matrix <- data.frame(
  "A" = c(0, 0, 0.3, 0),
  "B" = c(0.5, 0, 0, 0.6),
  "C" = c(0, 0, 0, 0),
  "D" = c(1, 0, 1, 0)
)
get_edgelist_from_adj_matrix(adj_matrix)
}
\seealso{
Other utility: 
\code{\link{check_if_local_machine_has_access_to_parallel_processing_functionalities}()},
\code{\link{check_if_local_machine_has_access_to_show_progress_functionalities}()},
\code{\link{get_adj_matrices_input_type}()},
\code{\link{get_inferences}()},
\code{\link{get_node_IDs_from_input}()},
\code{\link{standardize_adj_matrices}()}
}
\concept{utility}
