% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_inferences}
\alias{get_inferences}
\title{Get inferences from an fcmconfr output}
\usage{
get_inferences(fcmconfr_obj = list(), analysis = c("input", "aggregate", "mc"))
}
\arguments{
\item{fcmconfr_obj}{An fcmconfr output object}

\item{analysis}{The}
}
\value{
A dataframe (or list of dataframes) of inferences from the selected
analysis (analyses)
}
\description{
Given an fcmconfr output object, return the inferences of all or just a
specific analysis
}
\examples{
ex_conventional_fcmconfr <- fcmconfr(
  adj_matrices = sample_fcms$simple_fcms$conventional_fcms,
  # adj_matrices = group_conventional_fcms,
  # Aggregation and Monte Carlo Sampling
  aggregation_function = 'mean',
  monte_carlo_sampling_draws = 100,
  # Simulation
  # initial_state_vector = c(0, 0, 1, 0, 0, 0, 0, 0, 0),
  # clamping_vector = c(0, 0, 0, 0, 0, 0, 0, 0, 0),
  initial_state_vector = test_initial_state_vector,
  clamping_vector = test_clamping_vector,
  activation = 'kosko',
  squashing = 'tanh',
  lambda = 0.5,
  point_of_inference = "final",
  max_iter = 10000,
  min_error = 1e-05,
  # Inference Estimation (bootstrap)
  inference_estimation_function = "mean",
  inference_estimation_CI = 0.95,
  inference_estimation_bootstrap_reps = 1000,
  # Runtime Options
  show_progress = TRUE,
  parallel = TRUE,
  n_cores = 2,
  # Additional Options
  perform_aggregate_analysis = TRUE,
  perform_monte_carlo_analysis = TRUE,
  perform_monte_carlo_inference_bootstrap_analysis = TRUE,
  include_zero_weighted_edges_in_aggregation_and_mc_sampling = TRUE,
  include_monte_carlo_FCM_simulations_in_output = TRUE
)
get_inferences(ex_conventional_fcmconfr, analysis = c("input"))
get_inferences(ex_conventional_fcmconfr, analysis = c("input", "aggregate"))
get_inferences(ex_conventional_fcmconfr, analysis = c("input", "aggregate", "mc"))
get_inferences(ex_conventional_fcmconfr)


ex_ivfn_fcmconfr <- fcmconfr(
  adj_matrices = sample_fcms$simple_fcms$ivfn_fcms,
  # adj_matrices = group_ivfn_fcms,
  # Aggregation and Monte Carlo Sampling
  aggregation_function = 'mean',
  monte_carlo_sampling_draws = 100,
  # Simulation
  # initial_state_vector = c(0, 0, 1, 0, 0, 0, 0, 0, 0),
  # clamping_vector = c(0, 0, 0, 0, 0, 0, 0, 0, 0),
  initial_state_vector = test_initial_state_vector,
  clamping_vector = test_clamping_vector,
  activation = 'rescale',
  squashing = 'sigmoid',
  lambda = 1,
  point_of_inference = "final",
  max_iter = 1000,
  min_error = 1e-05,
  # Inference Estimation (bootstrap)
  inference_estimation_function = mean,
  inference_estimation_CI = 0.95,
  inference_estimation_bootstrap_reps = 1000,
  # Runtime Options
  show_progress = TRUE,
  parallel = TRUE,
  n_cores = 2,
  # Additional Options
  perform_aggregate_analysis = TRUE,
  perform_monte_carlo_analysis = TRUE,
  perform_monte_carlo_inference_bootstrap_analysis = TRUE,
  include_zero_weighted_edges_in_aggregation_and_mc_sampling = TRUE,
  include_monte_carlo_FCM_simulations_in_output = TRUE
)
get_inferences(ex_ivfn_fcmconfr, analysis = c("input"))
get_inferences(ex_ivfn_fcmconfr, analysis = c("input", "aggregate"))
get_inferences(ex_ivfn_fcmconfr, analysis = c("input", "aggregate", "mc"))
get_inferences(ex_ivfn_fcmconfr)


ex_tfn_fcmconfr <- fcmconfr(
  adj_matrices = sample_fcms$simple_fcms$tfn_fcms,
  # adj_matrices = group_tfn_fcms,
  # Aggregation and Monte Carlo Sampling
  aggregation_function = 'mean',
  monte_carlo_sampling_draws = 100,
  # Simulation
  initial_state_vector = c(1, 1, 1, 1, 1, 1, 1),
  clamping_vector = c(1, 0, 0, 0, 0, 0, 0),
  activation = 'rescale',
  squashing = 'sigmoid',
  lambda = 1,
  point_of_inference = "final",
  max_iter = 1000,
  min_error = 1e-05,
  # Inference Estimation (bootstrap)
  inference_estimation_function = mean,
  inference_estimation_CI = 0.95,
  inference_estimation_bootstrap_reps = 1000,
  # Runtime Options
  show_progress = TRUE,
  parallel = TRUE,
  n_cores = 2,
  # Additional Options
  perform_aggregate_analysis = TRUE,
  perform_monte_carlo_analysis = TRUE,
  perform_monte_carlo_inference_bootstrap_analysis = TRUE,
  include_zero_weighted_edges_in_aggregation_and_mc_sampling = TRUE,
  include_monte_carlo_FCM_simulations_in_output = TRUE
)
get_inferences(ex_tfn_fcmconfr, analysis = c("input"))
get_inferences(ex_tfn_fcmconfr, analysis = c("input", "aggregate"))
get_inferences(ex_tfn_fcmconfr, analysis = c("input", "aggregate", "mc"))
get_inferences(ex_tfn_fcmconfr)
}
\seealso{
Other utility: 
\code{\link{check_if_local_machine_has_access_to_parallel_processing_functionalities}()},
\code{\link{check_if_local_machine_has_access_to_show_progress_functionalities}()},
\code{\link{get_adj_matrices_input_type}()},
\code{\link{get_edgelist_from_adj_matrix}()},
\code{\link{get_node_IDs_from_input}()},
\code{\link{standardize_adj_matrices}()}
}
\concept{utility}
