% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{get_mc_simulations_inference_CIs_w_bootstrap}
\alias{get_mc_simulations_inference_CIs_w_bootstrap}
\title{get_mc_simulations_inference_CIs_w_bootstrap}
\usage{
get_mc_simulations_inference_CIs_w_bootstrap(
  mc_simulations_inference_df = data.frame(),
  confidence_interval = 0.95,
  bootstrap_reps = 1000,
  bootstrap_draws_per_rep = 1000,
  parallel = TRUE,
  n_cores = integer(),
  show_progress = TRUE
)
}
\arguments{
\item{mc_simulations_inference_df}{The final values of a set of fcm simulations; also the inference of a infer_fmcm object}

\item{confidence_interval}{What are of the distribution should be bounded by the
confidence intervals? (e.g. 0.95)}

\item{bootstrap_reps}{Repetitions for bootstrap process, if chosen}

\item{parallel}{TRUE/FALSE Whether to perform the function using parallel processing}

\item{n_cores}{Number of cores to use in parallel processing. If no input given,
will use all available cores in the machine.}

\item{show_progress}{TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply
from the pbapply package as the underlying function.}

\item{get_bootstrapped_means}{TRUE/FALSE Whether to perform bootstrap sampling to obtain
confidence intervals for the estimation of the mean value across simulations}

\item{bootstrap_samples_per_rep}{Number of samples to draw (with replacement) from
the data per bootstrap_rep}
}
\description{
This gets the mean of the distribution of simulated values
across a given iter. Also returns the bootstrapped mean of means of the
distribution of simulated values across a given iter if called.
}
\details{
This function is designed to streamline the process of getting the mean or bootstrapped
mean of means of a distribution of simulated values across individual iterations. Use get_bootstrapped_means
to estimate the confidence intervals for the mean value across simulations.
}
