% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_node_IDs_from_input}
\alias{get_node_IDs_from_input}
\title{get_node_IDs_from_input}
\usage{
get_node_IDs_from_input(adj_matrix = matrix())
}
\arguments{
\item{adj_matrix}{An n x n adjacency matrix that represents an FCM}
}
\description{
Determine node/concept IDs vector depending on the input adjacency matrix and
IDs
}
\details{
Node/concept IDs can come from two sources: the column names of the input
adjacency matrix or an explicitly-defined IDs vector. If an IDs vector is
given, confirm that it is applicable to the input adjacency matrix. If no
IDs are available, use C1, C2, ..., Cn for IDs.

Intended for developer use only to improve package readability.
}
