% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcmconfr.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Get fcmconfr Object Plot Data}
\usage{
get_plot_data(fcmconfr_object, filter_limit = 0.01)
}
\arguments{
\item{fcmconfr_object}{A direct output of the \code{\link{fcmconfr}} function}

\item{filter_limit}{Remove concepts whose inferences do not exceed this value}
}
\value{
A list of fcmconfr output dataframes organized to streamline
functionality with ggplot
}
\description{
This function parses through an \code{\link{fcmconfr}} output to gather and
organize the analysis results into dataframes that are constructed to be
plugged directly into a ggplot2 pipeline
}
\details{
This function produces slightly different outputs for \code{\link{fcmconfr}}
outputs generated from conventional, ivfn, and tfn FCMs
}
\examples{
NULL
}
