% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R
\name{ivfn}
\alias{ivfn}
\title{ivfn S3 class}
\usage{
ivfn(lower = double(), upper = double())
}
\arguments{
\item{lower}{lower limit of an IVFN (the lower value must be less than or equal to the upper value)}

\item{upper}{upper limit of an IVFN (the upper value must be greater or equal to the lower value)}
}
\description{
This class is an organization scheme for Grey Numbers (See
____, XXXX for example).
}
\details{
Interval Value Fuzzy Numbers (IVFNs) represent intervals within a possibility space (typically
[0, 1] or [-1, 1]) that contains the true value of an object.

The IVFN class does not perform any operations on its input, rather
it checks whether the input follows the defining criteria of IVFNs
}
\examples{
ivfn(lower = 0, upper = 1)
}
