% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R,
%   R/[DEPRECATED]triangular_fuzzy_numbers.R
\name{make_adj_matrix_w_tfns}
\alias{make_adj_matrix_w_tfns}
\title{make_adj_matrix_w_tfns}
\usage{
make_adj_matrix_w_tfns(lower = matrix(), mode = matrix(), upper = matrix())

make_adj_matrix_w_tfns(lower = matrix(), mode = matrix(), upper = matrix())
}
\arguments{
\item{lower}{An n x n adjacency matrix that represents the lower limits of
edges in an FCM}

\item{mode}{An n x n adjacency matrix that represents the modes of edges in an FCM}

\item{upper}{An n x n adjacency matrix that represents the upper limits of
edges in an FCM}
}
\description{
This constructs a triangular adjacency matrix from an adjacency matrix of the lower
limits of edges in an FCM, an adjacency matrix of the averages or expected values
of edges, and an adjacency matrix of the upper limits of edges in an FCM.

This constructs a triangular adjacency matrix from an adjacency matrix of the lower
limits of edges in an FCM, an adjacency matrix of the averages or expected values
of edges, and an adjacency matrix of the upper limits of edges in an FCM.
}
\details{
The input adjacency matrices must square n x n matrices with the same dimensions.
The input can be either matrix, data.table, tibble, or data.table type objects,
but the output will always be a data.frame. This is for output readability.
data.table and tibble objects work logically, but their outputs require
additional steps to parse from the user's perspective.

If the input matrices have named columns, those names will be carried over
in the triangular adjacency matrix. Otherwise, generic node IDs will be used
(C1, C2, ... Cn).

The input adjacency matrices must square n x n matrices with the same dimensions.
The input can be either matrix, data.table, tibble, or data.table type objects,
but the output will always be a data.frame. This is for output readability.
data.table and tibble objects work logically, but their outputs require
additional steps to parse from the user's perspective.

If the input matrices have named columns, those names will be carried over
in the triangular adjacency matrix. Otherwise, generic node IDs will be used
(C1, C2, ... Cn).
}
\examples{
make_adj_matrix_w_tfns(
 lower = matrix(data = c(0, 0.2, 0, 0.5), nrow = 2, ncol = 2),
 mode = matrix(data = c(0, 0.3, 0, 0.6), nrow = 2, ncol = 2),
 upper = matrix(data = c(0, 0.4, 0, 0.7), nrow = 2, ncol = 2)
)
make_adj_matrix_w_tfns(
 lower = matrix(data = c(0, 0.2, 0, 0.5), nrow = 2, ncol = 2),
 mode = matrix(data = c(0, 0.3, 0, 0.6), nrow = 2, ncol = 2),
 upper = matrix(data = c(0, 0.4, 0, 0.7), nrow = 2, ncol = 2)
)
}
