% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_model_generation_and_simulation.R
\name{monte_carlo_bootstrap_checks}
\alias{monte_carlo_bootstrap_checks}
\title{Check inputs for monte carlo bootstrap analysis}
\usage{
monte_carlo_bootstrap_checks(
  inference_function,
  confidence_interval,
  bootstrap_reps,
  bootstrap_draws_per_rep
)
}
\arguments{
\item{inference_function}{Estimate confidence intervals about the "mean" or "median" of
inferences from the monte carlo simulations}

\item{confidence_interval}{What are of the distribution should be bounded by the
confidence intervals? (e.g. 0.95)}

\item{bootstrap_reps}{Repetitions for bootstrap process, if chosen}

\item{bootstrap_draws_per_rep}{Number of samples to draw (with replacement) from
the data per bootstrap_rep}
}
\description{
Check inputs for monte carlo bootstrap analysis
}
\examples{
NULL
}
\seealso{
Other monte-carlo-model-generation-and-simulation: 
\code{\link{build_monte_carlo_fcms}()},
\code{\link{build_monte_carlo_fcms_from_conventional_adj_matrices}()},
\code{\link{build_monte_carlo_fcms_from_fuzzy_set_adj_matrices}()},
\code{\link{get_mc_simulations_inference_CIs_w_bootstrap}()},
\code{\link{infer_monte_carlo_fcm_set}()}
}
\concept{monte-carlo-model-generation-and-simulation}
