% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmconfr.R
\name{organize_fcmconfr_output}
\alias{organize_fcmconfr_output}
\title{organize_fcmconfr_output}
\usage{
organize_fcmconfr_output(...)
}
\arguments{
\item{...}{additional inputs; typically environmental variables}
}
\description{
This arranges fcmconfr inputs and outputs into a neatly-arranged list of lists
}
\details{
A standardized way to organize the fcmconfr into easily navigable data structures.

Use vignette("fcmconfr-class") for more information.
}
