% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcmconfr.R
\name{plot.fcmconfr}
\alias{plot.fcmconfr}
\title{Plot fcmconfr}
\usage{
\method{plot}{fcmconfr}(x, ...)
}
\arguments{
\item{x}{A direct output of the \code{\link{fcmconfr}} function}

\item{...}{Additional inputs:
- interactive Open plot in interactive shiny app
# Plot Formatting Parameters
- filter_limit Remove concepts whose inferences do not exceed this value
- coord_flip Swap x- and y-axes (i.e. rotate plot)
- text_font_size (Mainly for shiny use, i.e. leave blank) Set text font size
# Plot Aesthetic Parameters (and defaults)
- mc_avg_and_CIs_color = "blue"
- mc_inferences_color = "blue"
- mc_inferences_shape = 3
- ind_inferences_color = "black"
- ind_inferences_shape = 16
- agg_inferences_color = "red"
- agg_inferences_shape = 17
- ind_ivfn_and_tfn_linewidth = 0.1
- agg_ivfn_and_tfn_linewidth = 0.6}
}
\value{
A plot of an fcmconfr object's results
}
\description{
Print the output of the autoplot function (\code{\link{autoplot.fcmconfr}})
}
\examples{

# Example using TFN FCMs fcmconfr
tfn_example_fcmconfr <- fcmconfr(
  adj_matrices = sample_fcms$simple_fcms$tfn_fcms,
  # adj_matrices = group_tfn_fcms,
  # Aggregation and Monte Carlo Sampling
  agg_function = 'mean',
  num_mc_fcms = 100,
  # Simulation
  initial_state_vector = c(1, 1, 1, 1, 1, 1, 1),
  clamping_vector = c(1, 0, 0, 0, 0, 0, 0),
  activation = 'rescale',
  squashing = 'sigmoid',
  lambda = 1,
  point_of_inference = "final",
  max_iter = 1000,
  min_error = 1e-05,
  # Inference Estimation (bootstrap)
  ci_centering_function = mean,
  confidence_interval = 0.95,
  num_ci_bootstraps = 1000,
  # Runtime Options
  show_progress = TRUE,
  parallel = TRUE,
  n_cores = 2,
  # Additional Options
  run_agg_calcs = TRUE,
  run_mc_calcs = TRUE,
  run_ci_calcs = TRUE,
  include_zeroes_in_sampling = TRUE,
  mc_sims_in_output = TRUE
)


# Plot Defaults
plot(tfn_example_fcmconfr,
     interactive = FALSE, # Set to TRUE to open shiny app
     # Plot Formatting Parameters
     filter_limit = 1e-4,
     xlim = c(-1, 1),
     coord_flip = FALSE,
     # Plot Aesthetic Parameters
     mc_avg_and_CIs_color = "blue",
     mc_inferences_color = "blue",
     mc_inferences_alpha = 1,
     mc_inferences_shape = 3,
     ind_inferences_color = "black",
     ind_inferences_alpha = 1,
     ind_inferences_shape = 16,
     agg_inferences_color = "red",
     agg_inferences_alpha = 1,
     agg_inferences_shape = 17,
     ind_ivfn_and_tfn_linewidth = 0.1,
     agg_ivfn_and_tfn_linewidth = 0.6
)

# Changed from Plot Defaults
plot(tfn_example_fcmconfr,
     interactive = FALSE, # Set to TRUE to open shiny app
     # Plot Formatting Parameters
     filter_limit = 1e-4,
     xlim = c(-0.6, 0.6),
     coord_flip = FALSE,
     # Plot Aesthetic Parameters
     mc_avg_and_CIs_color = "blue",
     mc_inferences_color = "blue",
     mc_inferences_shape = 3,
     ind_inferences_color = "black",
     ind_inferences_shape = 16,
     agg_inferences_color = "red",
     agg_inferences_shape = 17,
     ind_ivfn_and_tfn_linewidth = 0.1,
     agg_ivfn_and_tfn_linewidth = 0.6
)


# Plot Defaults w/ Shiny App
# plot(tfn_example_fcmconfr,
#      interactive = TRUE, # Set to TRUE to open shiny app
#      # Plot Formatting Parameters
#      filter_limit = 1e-4,
#      coord_flip = FALSE,
#      text_font_size = 12,
#      # Plot Aesthetic Parameters
#      mc_avg_and_CIs_color = "blue",
#      mc_inferences_color = "blue",
#      mc_inferences_shape = 3,
#      ind_inferences_color = "black",
#      ind_inferences_shape = 16,
#      agg_inferences_color = "red",
#      agg_inferences_shape = 17,
#      ind_ivfn_and_tfn_linewidth = 0.1,
#      agg_ivfn_and_tfn_linewidth = 0.6
# )
}
