% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcmconfr.R
\name{plot.fcmconfr}
\alias{plot.fcmconfr}
\title{Plot fcmconfr}
\usage{
\method{plot}{fcmconfr}(x, ...)
}
\arguments{
\item{x}{A direct output of the \code{\link{fcmconfr}} function}

\item{...}{Additional inputs:
- filter_limit Remove concepts whose inferences do not exceed this value
- coord_flip Swap x- and y-axes (i.e. rotate plot)}
}
\description{
Print the output of the autoplot function (\code{\link{autoplot.fcmconfr}})
}
