% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmconfr.R
\name{print.fcmconfr}
\alias{print.fcmconfr}
\title{print.fcmconfr}
\usage{
\method{print}{fcmconfr}(x, ...)
}
\arguments{
\item{x}{an fcmconfr object}

\item{...}{additional inputs}
}
\value{
A console printout of fcmconfr results
}
\description{
This improves the readability of the fcmconfr output
}
\details{
Show the objects listed in the fcmconfr output \code{$inference} and \code{$params},
as well as \code{$bootstrap} if present in output. Additionally, this prints
descriptions/summaries of objects within each sub-list like inference_opts,
bootstrap_input_opts, etc.
}
\examples{
NULL
}
