% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmconfr_gui.R
\name{print.fcmconfr_input}
\alias{print.fcmconfr_input}
\title{Print fcmconfr_input}
\usage{
\method{print}{fcmconfr_input}(x, ...)
}
\arguments{
\item{x}{an fcmconfr object}

\item{...}{additional inputs}
}
\value{
A pasteable output to the console of a call to fcmconfr with the
selected variables
}
\description{
Print fcmconfr_input
}
\examples{
NULL
}
