% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_and_simulate_fcm.R
\name{print.infer_ivfn_or_tfn_fcm}
\alias{print.infer_ivfn_or_tfn_fcm}
\title{Print method for infer_ivfn_or_tfn_fcm objects}
\usage{
\method{print}{infer_ivfn_or_tfn_fcm}(x, ...)
}
\description{
Print method for infer_ivfn_or_tfn_fcm objects
}
