% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R
\name{print.ivfn}
\alias{print.ivfn}
\title{Print an Interval-Valued Fuzzy Number (IVFN) - S3 Class}
\usage{
\method{print}{ivfn}(x, ...)
}
\arguments{
\item{x}{a ivfn object}

\item{...}{additional inputs}
}
\description{
This prints an ivfn object
}
\seealso{
Other interval-valued-fuzzy-numbers: 
\code{\link{c.ivfn}()},
\code{\link{ivfn}()},
\code{\link{make_adj_matrix_w_ivfns}()}
}
\concept{interval-valued-fuzzy-numbers}
