% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_grey_cognitive_maps.R
\name{print.ivfn}
\alias{print.ivfn}
\title{print.ivfn}
\usage{
\method{print}{ivfn}(x, ...)
}
\arguments{
\item{x}{a grey_number object}

\item{...}{additional inputs}
}
\description{
This improves the readability of the output
}
\details{
Grey Numbers represent intervals within a possibility space (typically
[0, 1] or [-1, 1]) that contains the true value of an object.

The grey_number class does not perform any operations on its input, rather
it checks whether the input follows the defining criteria of grey numbers

grey_numbers are constructed using vctrs::new_vctr instead of the typical call to
structure() because ...

Use vignette("fgcm-class") for more information.
}
\examples{
grey_number(lower = 0, upper = 1)
}
