% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R
\name{print.tfn}
\alias{print.tfn}
\title{Print a Triangular Fuzzy Number (TFN)}
\usage{
\method{print}{tfn}(x, ...)
}
\arguments{
\item{x}{a tfn object}

\item{...}{additional inputs}
}
\description{
This improves the readability of the output

This prints a tfn object
}
\seealso{
Other triangular-fuzzy-numbers: 
\code{\link{make_adj_matrix_w_tfns}()},
\code{\link{tfn}()}
}
\concept{triangular-fuzzy-numbers}
