% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_triangular_cognitive_maps.R
\name{print.tfn}
\alias{print.tfn}
\title{print.tfn}
\usage{
\method{print}{tfn}(x, ...)
}
\arguments{
\item{x}{a tfn object}

\item{...}{additional inputs}
}
\description{
This improves the readability of the output
}
\details{
triangular Numbers represent intervals within a possibility space (typically
[0, 1] or [-1, 1]) that contains the true value of an object.

The tfn class does not perform any operations on its input, rather
it checks whether the input follows the defining criteria of triangular numbers

tfns are constructed using vctrs::new_vctr instead of the typical call to
structure() because ...

Use vignette("fcm_w_fcm_w_tfn-class") for more information.
}
\examples{
tfn(lower = 0, upper = 1)
}
