% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFNs_and_TFNs.R
\name{rtri}
\alias{rtri}
\title{rtri}
\usage{
rtri(n = integer(), lower = double(), mode = double(), upper = double())
}
\arguments{
\item{n}{number of samples to draw from the triangular distribution}

\item{lower}{lower limit or minimum of the sample space}

\item{mode}{peak of the sample space}

\item{upper}{upper limit or maximum of the sample space}
}
\description{
This pulls n samples from a triangular distribution described by shape parameters
defined by a lower limit, upper limit, and mode
}
\details{
Use vignette("fcmcmr-class") for more information.
}
