% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_fcms}
\alias{sample_fcms}
\title{sample_fcms}
\format{
## `salinization_conventional_fcms`
A list of conventional fcm adjacency matrices
}
\usage{
sample_fcms
}
\description{
An example fcm dataset collected by collecting stakeholder perceptions of
rising salinization in a socio-ecological system. To reduce file size and
protect stakeholder anonymity, the provided adjacency matrices contain a
subset of nodes and edges across the data used in the cited paper.
}
\details{
Given in Rippy et al, (In Review)
}
\keyword{datasets}
