% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{standardize_adj_matrices}
\alias{standardize_adj_matrices}
\title{standardize_adj_matrices}
\usage{
standardize_adj_matrices(adj_matrices = list(matrix()))
}
\arguments{
\item{adj_matrices}{A list of adj. matrix objects}
}
\description{
Given a list of adj. matrices of different sizes, transforms the inputs into
adj. matrices of the same size with 0 values added where edges are undefined
or not included in maps.

Solves the problem where Map 1 has nodes A and B but Map 2 has nodes B and C
and need to analyse them both together.
}
